==> analysis/01_save_SI_maps.Rmd <==
---
title: "Save maps a multitude of ways for SI materials"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary

Here we save supplementary maps. We save mean, standard deviation, number of species impacted, coefficient of variation, and proportion of species impacted maps for:
 - across all species and materials
 - by taxon across materials
 - by raw material across taxon

## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(data.table)
library(terra)
library(rnaturalearth)
library(strex)
library(janitor)
library(sf)
library(tools)
library(patchwork)
library(cowplot)
library(glue)
library(RColorBrewer)
library(ggpubr)
library(smoothr)
library(ggtext)
library(ggnewscale)
library(doParallel)
library(foreach)
library(qs)

source(here("src/directories.R"))

source(here("src/spatial.R"))

setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # set working directory to where this script is located
this_dir <- getwd()

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

impact_maps_dir <- file.path(biodiv_dir, "output")

globe_bbox <- rbind(c(-180, -90), c(-180, 90), 
                      c(180, 90), c(180, -90), c(-180, -90)) 

globe_border <- st_polygon(list(globe_bbox)) %>%
  st_sfc(crs = 4326) %>%
  st_sf(data.frame(rgn = 'globe', geom = .)) %>%
  smoothr::densify(max_distance = 0.5) %>%
  st_transform(crs = crs(moll_template))


```

```{r establish color palettes and themes}

## establish color palette from food footprint project
d_brown <- "#515256"
m_brown <- "#B27B54"
l_brown <- "#BC995F"
green <- "#A8B072"
yellow<- "#EFCE71"
l_yellow <- "#F7F6C1"

jv_pal <- c(l_yellow, yellow, green, l_brown, m_brown, d_brown)

continuous_pal <-  colorRampPalette(jv_pal, space="Lab", bias = 3.5)(10000)
continuous_pal_red <- c(continuous_pal, "#B90000")

light_gray <- "#F8F9FA"
final_palette <- c(light_gray, continuous_pal)

palette_red <- c(final_palette, "#B90000")

countries_shp <- ne_countries(scale = 110, returnclass = "sf") |> 
  st_transform(crs(moll_template)) # read in countries shapefile

theme_ohara <- function(base_size = 9) {
  theme(axis.ticks = element_blank(),
        text             = element_text(family = 'Helvetica',
                                        color = 'gray30', size = base_size),
        plot.title       = element_text(size = rel(1.25), hjust = 0, face = 'bold'),
        panel.background = element_blank(),
        legend.position  = 'right',
        panel.border     = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(colour = 'grey90', size = .25),
        legend.key       = element_rect(colour = NA, fill = NA),
        axis.line        = element_blank())
}


```

Read in marine and terrestrial dfs and combine

```{r}
marine_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds")) %>%
  filter(sensitivity_scenario == "original") %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area) 


terrestrial_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_terrestrial_impact.rds")) %>% 
    filter(sensitivity_scenario == "original") %>%

    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

all_overlap <- rbind(marine_overlap, terrestrial_overlap)

hab_areas <- all_overlap %>% 
  distinct(sciname, spp_type, total_hab_area)

saveRDS(hab_areas, here("analysis/int/all_spp_hab_areas.rds")) 

all_overlap <- all_overlap %>%
  dplyr::select(-total_hab_area)

```

Maps!

## Save mean, sd, and nspp maps 

Global impact maps across ingredients and taxon
 - dark red areas indicate >99th quantile of impacts
 - save a map for each scenario

```{r}
## make plots showing impacts for each scenario 

allocation_types <- c("mass", "economic", "ge")  
calc_types <- c("mean_prop", "nspp", "sd_prop", "mean_ext_risk", "rarity_mean_prop", "total_km2")
calc_types <- c("nspp", "total_km2")
types_df <- expand.grid(allocation_types, calc_types) %>%
  unite(new_var, Var1:Var2, sep = "_") %>%
  pull(new_var) %>%
  unique()

diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular") # we only do regular now

num_cores <- 12
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(type = types_df,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
  
  # type = "economic_total_km2"
  
 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{type}.qs"))

loop_df <-  qs::qread(filepath) 

if(str_detect(type, "mean")){
  
  if(str_detect(type, "mean_ext_risk")){
        fill_name = "Average extinction\nRisk"
        
  }else if(str_detect(type, "rarity_mean_prop")){
  fill_name = "Average rarity\nweighted\nproportion\nhabitat\nimpacted"

  }else{
    
      fill_name = "Average proportion\nhabitat\nimpacted"

  }
  
  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){
 
# diet_type = "plant-dominant"
# fcr = "regular"
      
  loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)
  
quantile_value <- quantile(loop_loop_df$value, probs = 0.99, na.rm = TRUE)
  
max_val =  formatC(max(loop_loop_df$value, na.rm = TRUE), format = "e",  digits = 2)
quant_val = formatC(quantile_value, format = "e", digits = 2)

  loop_loop_df <- loop_loop_df %>% 
    mutate(value_new = ifelse(value > quantile_value, quantile_value, value))
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value_new, color = value_new)) +
     geom_tile(fill = ifelse(loop_loop_df$value_new < quantile_value, "#B90000", NA)) + 
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white", labels = scales::label_scientific()) +
            scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}"),
       caption = glue("99th quantile range (red areas): {quant_val} - {max_val}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/across_taxon_ingredient/{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

}else{
  
 # type = "ge_sd_prop"
  
 if(str_detect(type, "nspp")){
    fill_name = "Number of\nspecies\nimpacted"
  }else if(str_detect(type, "total_km2")){

fill_name <- expression("Total km"^2)  
  }else {
       fill_name = "sdev"
  }
  
    plot_list <- list()
  
    for(diet_type in diets){
    for(fcr in fcrs){
  
loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)

  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value, color = value)) +
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal, na.value = "white") +
            scale_color_gradientn(colors = continuous_pal, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + 
  labs(fill = fill_name,
       title = glue("{diet_type}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot
    }
    }
  
ggsave(glue(here(this_dir, "SI_plots/across_taxon_ingredient/{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 1, ncol = 2), width = 10, height = 8, bg = "white")

}

}


stopCluster(cl)

```

Now save SI fig 2 and 4

> SI Figure 2. The number of species impacted in each cell for each feed formulation scenario. White areas represent areas with no impacted habitat or production of materials. Related to Figures 1 and 2. 

> SI Figure 4. The total km2 of impacted habitat in each cell. White areas represent areas with no impacted habitat or production of materials. Related to Figure 1. 

```{r}
file.copy(
  from = here::here("analysis/SI_plots/across_taxon_ingredient/economic_total_km2_plot.png"),
  to   = here::here("analysis/SI_plots/SI_Fig_4_total_km2.png"),
  overwrite = TRUE
)

file.copy(
  from = here::here("analysis/SI_plots/across_taxon_ingredient/economic_nspp_plot.png"),
  to   = here::here("analysis/SI_plots/SI_Fig_2_nspp_impacted.png"),
  overwrite = TRUE
)
```

Global maps across ingredients per taxon 

```{r}

diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular") # only include regular now
allocations <- c("economic", "ge", "mass")
taxon <- unique(all_overlap$spp_type)
calc_types <- c("mean_prop", "mean_ext_risk", "sd_prop", "nspp", "total_km2", "rarity_mean_prop")
     
  
# make and save plots across ingredients per taxon  
num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)

foreach(allocation_type = allocations,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here', 'scales')) %dopar% {
    for(taxa_type in taxon){
      for(type in calc_types){

# allocation_type = "economic"
# taxa_type = "Bird"
# type = "rarity_mean_prop"

        loop_df <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{type}.qs"))) %>%
          filter(!is.na(value))

if(str_detect(type, "mean")){
  
  if(str_detect(type, "mean_ext_risk")){
        fill_name = "Average extinction\nRisk"
        
  }else if(str_detect(type, "rarity_mean_prop")){
  fill_name = "Average rarity\nweighted\nproportion\nhabitat\nimpacted"

  }else{
    
      fill_name = "Average proportion\nhabitat\nimpacted"

  }
  
  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){

      # diet_type = "plant-dominant"
      # fcr = "regular"
      
  loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)
  
quantile_value <- quantile(loop_loop_df$value, probs = 0.99, na.rm = TRUE)

max_val =  formatC(max(loop_loop_df$value), format = "e", digits = 2)
quant_val = formatC(quantile_value, format = "e", digits = 2)
  
  loop_loop_df <- loop_loop_df %>% 
    mutate(value_new = ifelse(value > quantile_value, quantile_value, value))
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value_new, color = value_new)) +
     geom_tile(fill = ifelse(loop_loop_df$value_new < quantile_value, "#B90000", NA)) + 
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white", labels = scales::label_scientific()) +
            scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}"), 
        caption = glue("99th quantile range (red areas): {quant_val} - {max_val}")) + 
     guides(color = "none") 
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/by_taxa/{allocation_type}_{taxa_type}_{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

}else{
  
 # type = "ge_total_km2"
  
 if(str_detect(type, "nspp")){
    fill_name = "Number of\nspecies\nimpacted"
  }else if(str_detect(type, "total_km2")){

fill_name <- expression("Total km"^2)  
  }else {
       fill_name = "sdev"
  }
  
  
    plot_list <- list()
  
    for(diet_type in diets){
    for(fcr in fcrs){
  
loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)

  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value, color = value)) +
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal, na.value = "white") +
            scale_color_gradientn(colors = continuous_pal, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot
    }
    }
    
    ggsave(glue(here(this_dir, "SI_plots/by_taxa/{allocation_type}_{taxa_type}_{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

    }
  }
    }
}

stopCluster(cl)

```

Global maps by raw material across taxon

```{r}

allocations <- c("economic", "ge", "mass")
calc_types <- c("mean_ext_risk", "nspp", "sd_prop", "mean_prop", "total_km2", "rarity_mean_prop")
raw_mats <- c("Soybean", "Wheat", "Sunflower", "CropsNES", "Rapeseed", "Maize", "forage fish", "trimmings fish", "Pulses")

num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)

foreach(allocation_type = allocations,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
    for(raw_mat in raw_mats){
      for(type in calc_types){
        
      # raw_mat = "Soybean"
      # allocation_type = "economic"
      # type = "sd_prop"
        
                loop_df <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_material/{allocation_type}_{raw_mat}_{type}.qs"))) 
 
diets <- unique(loop_df$diet)
  

if(str_detect(type, "mean")){
  
  if(str_detect(type, "mean_ext_risk")){
        fill_name = "Average extinction\nRisk"
        
  }else if(str_detect(type, "rarity_mean_prop")){
  fill_name = "Average rarity\nweighted\nproportion\nhabitat\nimpacted"

  }else{
    
      fill_name = "Average proportion\nhabitat\nimpacted"

  }
  
  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){

      # diet_type = "plant-dominant"
      # fcr = "regular"
      
  loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)
  
quantile_value <- quantile(loop_loop_df$value, probs = 0.99, na.rm = TRUE)
  
max_val =  formatC(max(loop_loop_df$value), digits = 2, format = "e")
quant_val = formatC(quantile_value, digits = 2, format = "e")

  loop_loop_df <- loop_loop_df %>% 
    mutate(value_new = ifelse(value > quantile_value, quantile_value, value))
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value_new, color = value_new)) +
     geom_tile(fill = ifelse(loop_loop_df$value_new < quantile_value, "#B90000", NA)) + 
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white", labels = scales::label_scientific()) +
            scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}"),
       caption = glue("99th quantile range (red areas): {quant_val} - {max_val}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/by_material/{allocation_type}_{raw_mat}_{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

}else{
  
 # type = "ge_sd_prop"
  
 if(str_detect(type, "nspp")){
   
    fill_name = "Number of\nspecies\nimpacted"
  }else if(str_detect(type, "total_km2")){
    
fill_name <- expression("Total km"^2)  
  }else {
    
       fill_name = "sdev"
  }
  
    plot_list <- list()
  
    for(diet_type in diets){
    for(fcr in fcrs){
  
loop_loop_df <- loop_df %>%
    filter(diet == diet_type, 
           fcr_type == fcr)

  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = value, color = value)) +
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal, na.value = "white") +
            scale_color_gradientn(colors = continuous_pal, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot
    }
    }
    
    ggsave(glue(here(this_dir, "SI_plots/by_material/{allocation_type}_{raw_mat}_{type}_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

    }
  }
    }
}

stopCluster(cl)

```

## Make coefficient of variation maps 

Calculate and save coefficient of variation maps by each of the above categories (across taxa and material), (by taxa, across material), (across taxa by material)

Across taxa and materials 

```{r}

allocation_types <- c("mass", "economic", "ge")  
diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular")

num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(allocation_type = allocation_types,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
  
  # allocation_type = "economic"
  
 filepath_mean <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{allocation_type}_mean_prop.qs"))
 
  filepath_sd <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{allocation_type}_sd_prop.qs"))

loop_df_mean <-  qs::qread(filepath_mean) %>%
  dplyr::select(-calc_type) %>%
  rename(mean_prop = value)

loop_df_sd <-  qs::qread(filepath_sd) %>%
    dplyr::select(-calc_type) %>%
  rename(sd_prop = value)

loop_df_all <- loop_df_mean %>%
  left_join(loop_df_sd) %>%
  mutate(cv = sd_prop/mean_prop)

  
  fill_name = "Coefficient\nof variation"
  
  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){

      # diet_type = "plant-dominant"
      # fcr = "regular"
      
  loop_loop_df <- loop_df_all %>%
    filter(diet == diet_type, 
           fcr_type == fcr)
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = cv, color = cv)) +
      geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white") +
            scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/across_taxon_ingredient/{allocation_type}_cv_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

}

stopCluster(cl)

```


By taxa across materials

```{r}
diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular")
allocations <- c("economic", "ge", "mass")
taxon <- unique(all_overlap$spp_type)

num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(allocation_type = allocations,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
    for(taxa_type in taxon){

# allocation_type = "economic"
# taxa_type = "Bird"

  loop_df_mean <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_mean_prop.qs"))) %>%
      dplyr::select(-calc_type) %>%
  rename(mean_prop = value)
        
  loop_df_sd <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_sd_prop.qs"))) %>%
      dplyr::select(-calc_type) %>%
  rename(sd_prop = value)
  
  
  loop_df_all <- loop_df_mean %>%
    left_join(loop_df_sd) %>%
    mutate(cv = sd_prop/mean_prop)

  
    fill_name = "Coefficient\nof variation"

  
  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){

      # diet_type = "plant-dominant"
      # fcr = "efficient"
      
  loop_loop_df <- loop_df_all %>%
    filter(diet == diet_type, 
           fcr_type == fcr)
  
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = cv, color = cv)) +
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white") +
        scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}")) + 
     guides(color = "none") 
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/by_taxa/{allocation_type}_{taxa_type}_cv_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")

    }
        }
stopCluster(cl)



```


Global maps by raw material across taxon

```{r}

allocations <- c("economic", "ge", "mass")
raw_mats <- c("Soybean", "Wheat", "Sunflower", "CropsNES", "Rapeseed", "Maize", "forage fish", "trimmings fish", "Pulses")

num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)

foreach(allocation_type = allocations,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
    for(raw_mat in raw_mats){
        
      # raw_mat = "Soybean"
      # allocation_type = "economic"
      # type = "sd_prop"
        
                loop_df_mean <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_material/{allocation_type}_{raw_mat}_mean_prop.qs"))) %>%
                    dplyr::select(-calc_type) %>%
  rename(mean_prop = value)
                loop_df_sd <- qread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_material/{allocation_type}_{raw_mat}_sd_prop.qs"))) %>%
                    dplyr::select(-calc_type) %>%
  rename(sd_prop = value)
                
            loop_df_all <- loop_df_mean %>%
              left_join(loop_df_sd) %>%
              mutate(cv = sd_prop/mean_prop)
 
diets <- unique(loop_df_all$diet)
  
  fill_name = "Coefficient\nof variation"

  plot_list <- list()
  
  for(diet_type in diets){
    for(fcr in fcrs){

      # diet_type = "plant-dominant"
      # fcr = "efficient"
      
  loop_loop_df <- loop_df_all %>%
    filter(diet == diet_type, 
           fcr_type == fcr)

  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = cv, color = cv)) +
            geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
       scale_fill_gradientn(colors = continuous_pal_red, na.value = "white") +
            scale_color_gradientn(colors = continuous_pal_red, na.value = "white") +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}, {fcr}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

    }
  } 

ggsave(glue(here(this_dir, "SI_plots/by_material/{allocation_type}_{raw_mat}_cv_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 2), width = 10, height = 8, bg = "white")


  }
}


stopCluster(cl)

```

## Proportion of species impacted maps

Here we will make maps which describe the proportion of species which are impacted: 

nspp impacted / total species richness map

For each diet scenario (just filter for regular, as the number of species impacted does not change between fcr scenarios, only the magnitude)

```{r}

allocation_types <- c("mass", "economic", "ge")  
diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular")

richness_files <- list.files(here("data/spatial/output"), pattern = "richness", full.names = TRUE) 

richness_stack <- rast(richness_files) %>%
  app(., "sum", na.rm = TRUE) %>%
  as.data.frame(., xy = TRUE, na.rm = FALSE) %>%
  rename(total_nspp = sum) %>%
  mutate(total_nspp = ifelse(is.na(total_nspp), 0, total_nspp))


richness_plot <- rast(richness_files) %>%
  app(., "sum", na.rm = TRUE)

test <- sum(richness_plot[] == 0 | is.na(richness_plot[])) # 1613739
test <- sum(richness_plot[] == 0, na.rm = TRUE) # 0

test <- sum(is.na(richness_plot[])) # 1613739
# so when joining, we should have 6561238 - 1613739 = 4947499 cells


num_cores <- 3
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(allocation_type = allocation_types,
        .packages = c('stringr', 'qs', 'tidyverse', 'tools', 'strex', 'glue', 'cowplot', 'here')) %dopar% {
  
  # allocation_type = "economic"
  
 filepath_nspp <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{allocation_type}_nspp.qs"))
 

loop_df_nspp <-  qs::qread(filepath_nspp) %>%
  dplyr::select(-calc_type) %>%
  rename(nspp_impacted = value) %>%
  filter(fcr_type == "regular")

  
  fill_name = "Proportion\nof species\nimpacted"
  
  plot_list <- list()
  
  for(diet_type in diets){

# diet_type = "fish-dominant"
# fcr = "regular"
      
  loop_loop_df <- loop_df_nspp %>%
    filter(diet == diet_type) %>%
      right_join(richness_stack, by = c("long" = "x", "lat" = "y")) %>%
  mutate(nspp_impacted = ifelse(is.na(nspp_impacted), 0, nspp_impacted)) %>%
  mutate(prop_spp_impacted = nspp_impacted/total_nspp) %>%
    mutate(diet = diet_type) %>%
    filter(!is.na(prop_spp_impacted))
  
  loop_plot <-  ggplot() + 
      geom_tile(data = loop_loop_df, aes(x = long, y = lat, fill = prop_spp_impacted, color = prop_spp_impacted)) +
     geom_sf(data = countries_shp, fill = NA, colour = "grey") +
       scale_fill_gradientn(colors = continuous_pal_red) +
            scale_color_gradientn(colors = continuous_pal_red) +
            geom_sf(data = globe_border,
            fill = NA, color = 'grey70',
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
      theme(plot.title = element_text(hjust = 0.5), 
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.margin = margin(r = 10)) + # add a margin to legend text so it isn't cut off when saving 
  labs(fill = fill_name,
       title = glue("{diet_type}")) + 
     guides(color = "none")
        
      plot_list[[length(plot_list) + 1]] <- loop_plot

  } 

ggsave(glue(here(this_dir, "SI_plots/across_taxon_ingredient/{allocation_type}_prop_spp_plot.png")), plot = plot_grid(plotlist = plot_list, nrow = 1), width = 10, height = 8, bg = "white")

}

stopCluster(cl)


## for economic: Average proportion of species impacted across all cells is 0.31 (plant dominant) and 0.3 (fish dominant). Filtering out cells with no species impacted, average prop of species impacted is 0.68 (plant) and 0.68 (fish).

## ~78% of spp are impacted, 85% are vulnerable
```
