==> analysis/03_publication_plots.Rmd <==
---
title: "Publication plots: non maps"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary

This code produces figures 2 and 3, as well as some supplementary plots:

Figure 2: 
> Figure 2. Distribution of proportion of global habitat area impacted for species. Each point is a species. Only impacted species are shown in the plot. Orange points indicate mean across species within the taxa and scenario and orange lines indicate the median. Boxes represent the interquartile range (IQR, quartile Q1 to Q3); whiskers indicate observations 1.5x IQR below (above) Q1 (Q3) the box. The panel on the right shows points >95th percentile, and total number of assessed species per taxa. 


Figure 3: 
> Figure 3. Proportion of each taxa’s global habitat area which is impacted by raw material for each aquafeed scenario.


## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(data.table)
library(terra)
library(strex)
library(patchwork)
library(cowplot)
library(glue)
library(RColorBrewer)
library(ggpubr)
library(qs)
library(parallel)
library(doParallel)
library(janitor)

source(here("src/directories.R"))

source(here("src/spatial.R"))

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

impact_maps_dir <- file.path(biodiv_dir, "output")

this_dir <- here("analysis/")

```

```{r}
# Create a named vector with colors for each spp_type
color_scale <- readRDS(here("prep/03_prep_spp_habitats/int/color_scale.RDS"))

pal_mats <- c("Soybean" = "#328C41",
              "Trimmings fish" = "#8FBAED",
              "Sunflower" = "#4277A9",
              "Wheat" = "#FFE96D",
              "Rapeseed" = "#AC9900",
              "Pulses" = "#79E0DA",
              "Forage fish" = "#3DB1AB",
              "Maize" = "#0A6B66",
              "Other crops" = "#BC94C6")

theme_ohara <- function(base_size = 9) {
  theme(axis.ticks = element_blank(),
        text             = element_text(family = 'Helvetica',
                                        color = 'gray30', size = base_size),
        plot.title       = element_text(size = rel(1.25), hjust = 0, face = 'bold'),
        panel.background = element_blank(),
        legend.position  = 'right',
        panel.border     = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(colour = 'grey90', size = .25),
        legend.key       = element_rect(colour = NA, fill = NA),
        axis.line        = element_blank())
}

```

Read in marine and terrestrial dfs and combine

```{r}
marine_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds")) %>%
    filter(sensitivity_scenario == "original") %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

length(unique(marine_overlap$sciname)) # 23029

terrestrial_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_terrestrial_impact.rds")) %>% 
    filter(sensitivity_scenario == "original") %>%
      dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

all_overlap <- rbind(marine_overlap, terrestrial_overlap) %>%
    mutate(taxon = str_to_sentence(spp_type)) %>%
  mutate(taxon = case_when(
    taxon == "Terrestrial mammal" ~ "Terrestrial mammals",
    taxon == "Bird" ~ "Birds",
    taxon == "Fish" ~ "Finfish", 
    taxon == "Marine mammal" ~ "Marine mammals",
    taxon == "Marine plant" ~ "Marine plants",
    TRUE ~ taxon)) %>%
  mutate(ingredient = str_replace(ingredient, "CropsNES", "Other crops"))

length(unique(all_overlap$sciname)) # 54625

```

Some random stats for paper:

 - 54628 species assessed
 - 466 more spp impacted under plant-dominant
 - ~78% spp that are assessed are impacted
 - ~85% are vulnerable

```{r}
taxon_names <- all_overlap %>%
  distinct(spp_type, taxon)

spp_impacted <- all_overlap %>%
  filter(impact_total > 0) %>%
  group_by(taxon, diet, allocation, fcr_type) %>%
  summarise(spp_count_impacted = n_distinct(sciname)) %>%
  ungroup()

aquamaps_spp <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  distinct(spp_type = taxon, sciname = species)
  
eyres_spp <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(spp_type, sciname = scientific_name) %>%
  mutate(sciname = tolower(sciname)) %>%
  mutate(spp_type = case_when(
    spp_type == "mammals" ~ "Terrestrial mammal",
    spp_type == "amphibians" ~ "Amphibians", 
    spp_type == "reptiles" ~ "Reptiles", 
    spp_type == "birds" ~ "Bird"
  ))

spp_assessed <- rbind(aquamaps_spp, eyres_spp) %>%
  left_join(taxon_names) %>%
  group_by(taxon) %>%
  summarise(spp_count_assessed = n_distinct(sciname)) %>%
  ungroup() %>%
  filter(!is.na(taxon))

sum(spp_assessed$spp_count_assessed) # 54628

spp_assessed_impacted <- spp_impacted %>%
  left_join(spp_assessed) 

spp_impact_test <- spp_assessed_impacted %>%
  group_by(diet, allocation, fcr_type) %>%
  summarise(total_impact = sum(spp_count_impacted, na.rm = TRUE),
            total_assessed = sum(spp_count_assessed, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop = total_impact/total_assessed)

spp_total_hab <- all_overlap %>%
  distinct(sciname, spp_type, total_hab_area) %>%
  mutate(id = row_number())

```

**Figure 2: Mean proportion impacted**

 > Figure 2. Distribution of proportion of global habitat area impacted for species. Each point is a species. Only impacted species are shown in the plot. Orange points indicate mean across species within the taxa and scenario and orange lines indicate the median. Boxes represent the interquartile range (IQR, quartile Q1 to Q3); whiskers indicate observations 1.5x IQR below (above) Q1 (Q3) the box. The panel on the right shows points >95th percentile, and total number of assessed species per taxa. 

 - Look at ranges of values
 - Split into two plots, like in global food paper; 1 is <95th quantile, 1 is >95th quantile 
 - One point for each species with box and whisker chart overlaid
 - Add colored (orange) points for average within taxa
 
```{r}
spp_assessed_plot_df <- spp_assessed %>%
  mutate(spp_count_str = glue("n = {spp_count_assessed}"))

all_overlap_df <- all_overlap %>%
  rename(fcr = fcr_type) %>%
        separate_wider_delim(ingredient, delim = "_", names = c("raw_material", "ingredient")) %>% 
  mutate(raw_material = str_to_sentence(raw_material), 
         ingredient = str_to_sentence(ingredient),
         allocation = str_to_sentence(allocation),
         diet = str_to_sentence(diet),
         fcr = str_to_sentence(fcr)) %>%
    filter(allocation == "Economic",
           fcr == "Regular")

summary_df_spp <- all_overlap_df %>%
  group_by(sciname, diet, fcr, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(impact_total > 0) %>%
  left_join(spp_assessed_plot_df) %>%
  left_join(spp_total_hab %>% dplyr::select(-spp_type, -id)) %>%
  mutate(prop_impact = impact_total/total_hab_area)  %>%
  group_by(taxon, diet, fcr) %>%
  mutate(mean_prop = mean(prop_impact,na.rm=TRUE),
         median_prop = median(prop_impact,na.rm=TRUE)) %>%
  ungroup() %>%
  mutate(terr_marine = ifelse(taxon %in% c("Terrestrial mammals", "Birds", "Reptiles", "Amphibians"), "Terrestrial", "Marine"))


taxon_hab_area <- spp_total_hab %>%
  dplyr::select(-sciname, -id) %>%
  group_by(spp_type) %>%
  summarise(total_hab_area = sum(total_hab_area, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_names) %>%
  dplyr::select(-spp_type)
  

summary_df_all <- all_overlap_df %>%
  group_by(diet, fcr, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_hab_area) %>%
  mutate(prop_impact = impact_total/total_hab_area)

quant_95 <- quantile(summary_df_spp$prop_impact, 0.95)


summary_df_spp_out <- summary_df_spp %>%
  filter(prop_impact >= quant_95) %>%
  group_by(diet, fcr, allocation, taxon) %>%
  mutate(spp_outliers = n_distinct(sciname)) %>%
  ungroup() 

join_df <- summary_df_spp_out %>%
  distinct(taxon, diet, fcr, allocation, spp_outliers)

summary_df_spp_reg <- summary_df_spp %>%
  filter(prop_impact > quant_95) %>%
  left_join(join_df) %>%
  mutate(spp_impacted = spp_count_assessed - spp_outliers) %>%
  mutate(prop_outlier = spp_outliers/spp_count_assessed)

spp_assessed_reg_outlier <- distinct(summary_df_spp_reg, taxon, diet, spp_impacted, spp_outliers, prop_outlier) %>%
    add_row(taxon = "Corals", diet = "Plant-dominant", spp_impacted = 249, spp_outliers = 0, prop_outlier = 0) %>%
      add_row(taxon = "Marine plants", diet = "Plant-dominant", spp_impacted = 280, spp_outliers = 0, prop_outlier = 0) %>%
  arrange(taxon, diet) %>% 
  mutate(perc_outlier = round(prop_outlier, 3)*100) %>% 
  distinct(taxon, perc_outlier) %>%
group_by(taxon) %>%
  summarize(perc_outlier = paste0("(", paste(perc_outlier, collapse = "%, "), "%)"))


spp_assessed_reg <- distinct(summary_df_spp_reg, taxon, spp_impacted, spp_outliers) %>%
  group_by(taxon) %>%
  slice(1) %>%
  mutate(total_spp = spp_outliers + spp_impacted) %>%
  mutate(total_spp = format(total_spp, big.mark = ",")) %>%
  left_join(spp_assessed_reg_outlier) %>%
    mutate(spp_impacted_str = glue("n = {spp_impacted}"),
           spp_outliers_str = glue("n = {spp_outliers}"),
           spp_count_assessed = glue("n = {total_spp}")) %>%
      mutate(spp_count_assessed_perc = paste0(spp_count_assessed, ", ", perc_outlier))

# Set the levels of the taxon variable in summary_df_test
summary_df_spp$taxon <- with(summary_df_spp, reorder(taxon, prop_impact, FUN = function(x) mean(x)))

diet_cols = c("Fish-dominant" = "#0D6EB5", "Plant-dominant" = "#7BD67F")

## make boxplot of prop impacted

plot1 <- ggplot(summary_df_spp %>% filter(prop_impact > 0), aes(x = prop_impact)) +
  geom_point(aes(y = taxon, color = diet), alpha = 0.2, position = position_jitterdodge(dodge.width = -0.9)) + 
  geom_vline(xintercept = 0, color = "black") +
  geom_point(data = summary_df_spp, aes(x = mean_prop, y = taxon, group = diet),
             position = position_dodge(width = -0.9),
             shape = 21, color = "red", fill = "orange", size = 7) +
     geom_boxplot(aes(y = taxon, fill = diet), outlier.fill = NA, outlier.color = NA, position = position_dodge(width = -0.9), alpha = 0, colour = "orange", fatten = 5) +
   geom_boxplot(aes(y = taxon, fill = diet), outlier.fill = NA, outlier.color = NA, position = position_dodge(width = -0.9), alpha = 0, fatten = NULL) +
  theme_ohara() +
      labs(x = "Proportion of global habitat area impacted", fill = 'Diet', color = 'Diet') +
  scale_x_continuous(expand = c(0,0), limits = c(0, quant_95)) +
  scale_color_manual(values = diet_cols, breaks = names(diet_cols), guide = guide_legend(reverse = FALSE)) +
  theme(panel.background = element_blank(),
            axis.title.y = element_blank(),
            axis.text.x = element_text(size = 16, angle = 45, hjust = 1),
         axis.text.y = element_text(size = 16),
        strip.text = element_text(size = 16),
        axis.title = element_text(size = 16),
        legend.position = "none", 
        text = element_text(family = "Arial"))


# get legend

quant_95_round <- round(quant_95, 6)

plot2 <- ggplot(summary_df_spp %>% filter(prop_impact > quant_95), aes(x = prop_impact)) +
  geom_point(aes(y = taxon, color = diet), alpha = 0.4, position = position_jitterdodge(dodge.width = -0.9)) + 
  geom_vline(xintercept = quant_95, color = "black") +
  theme_ohara() +
      labs(x = "", y = "") +
  scale_x_continuous(expand = c(0,0), limits = c(quant_95, 0.018), breaks = c(quant_95, seq(0.003, 0.012, by = 0.003)), labels = c("0.000058", "0.003", "0.006", "0.009", "0.012")) +
  scale_color_manual(values = diet_cols, breaks = names(diet_cols), guide = guide_legend(reverse = FALSE)) +
  theme(panel.background = element_blank(),
            axis.title.y = element_blank(),
            axis.text.x = element_text(size = 16, angle = 45, hjust = 1),
         axis.text.y = element_blank(),
        strip.text = element_text(size = 16),
        axis.title = element_text(size = 16),
        legend.position = "none", 
        text = element_text(family = "arial")) +
      geom_text(data = spp_assessed_reg, aes(label = spp_count_assessed, y = taxon),
                x = 0.011, hjust = 0, color = 'black',
                family = 'arial',
                size = 6) 


plot1_legend <- get_legend(ggplot(summary_df_spp, aes(x = mean_prop)) +
  geom_boxplot(aes(y = taxon, fill = diet, color = diet), 
               width = 0.3, position = position_dodge(width = -0.7),
               size = 0.3, outlier.size = 0.3) + 
  theme_ohara() + 
    theme(panel.background = element_blank(),
            legend.text = element_text(size = 16),
            legend.title = element_text(size = 16),
            legend.key.size = unit(.20, 'cm'),
            legend.background = element_blank()) + 
    scale_fill_manual(values = diet_cols, breaks = names(diet_cols), guide = guide_legend(reverse = FALSE)) + 
  scale_color_manual(values = diet_cols, breaks = names(diet_cols), guide = guide_legend(reverse = FALSE)) + 
    labs(x = "Proportion of habitat area impacted", fill = 'Diet', color = 'Diet'))


# draw plots together 
p <- ggdraw() + draw_plot((plot1 + plot2 +  plot_layout(widths = c(3, 1)))) + draw_plot(plot1_legend,  x = .63, y = 0.15, height = .10, width = .15) 

ggsave(plot = p, here(this_dir, "MS_plots/fig2_economic_boxplot_prop_orange.png"), width = 16, height = 9, units = "in", bg = "white")

```
 

**Figure 3: Proportion of taxa global habitat area impacted**
> Figure 3. Proportion of each taxa’s global habitat area which is impacted by raw material for each aquafeed scenario.

```{r}

impacted_spp <- all_overlap_df %>%
  filter(impact_total > 0) %>%
  pull(sciname) %>%
  unique()

taxon_hab_area <- spp_total_hab %>%
  filter(sciname %in% impacted_spp) %>%
  dplyr::select(-sciname, -id) %>%
  group_by(spp_type) %>%
  summarise(total_hab_area = sum(total_hab_area, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_names) %>%
  dplyr::select(-spp_type)

summary_df_barcharts <- all_overlap_df %>%
  group_by(diet, fcr, raw_material, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(impact_total > 0) %>%
  left_join(taxon_hab_area) %>%
  mutate(prop_impact = impact_total/total_hab_area) %>%
    mutate(prop_impact_mirror = ifelse(diet == "Fish-dominant", -prop_impact, prop_impact))

## switch the raw materials and species in that plot

# Calculate the *total* mirrored impact per taxon (summed over raw materials)
sum_impacts <- summary_df_barcharts %>%
  group_by(diet, taxon) %>%
  summarize(total_impact = sum(prop_impact_mirror, na.rm = TRUE), .groups = "drop")

# Find the maximum absolute value
max_val <- max(abs(sum_impacts$total_impact), na.rm = TRUE)

summary_df_p <- summary_df_barcharts %>% filter(diet == "Plant-dominant")

summary_df_p$taxon <- with(summary_df_p, reorder(taxon, prop_impact, FUN = sum))

# 4. Plot
p_plot <- ggplot(summary_df_p, aes(x = prop_impact_mirror, 
                                 y = taxon, 
                                 fill = raw_material)) +
    geom_bar(stat = "identity") +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = NULL, 
         fill = "Raw material") +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.y = element_blank(),
        axis.text.x = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8),
        text = element_text(family = "arial"),
        axis.ticks.y = element_blank(),
        plot.margin = margin(0, 0, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(0, max_val),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
        breaks = c(0.00001, 0.00002, 0.00003)
    )

summary_df_f <- summary_df_barcharts %>% filter(diet == "Fish-dominant")

summary_df_f$taxon <- factor(summary_df_f$taxon, levels = levels(summary_df_p$taxon))

# 4. Plot
f_plot <- ggplot(summary_df_f, aes(x = prop_impact_mirror, 
                                 y = taxon, 
                                 fill = raw_material)) +
    geom_bar(stat = "identity") +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = "Taxa", 
         fill = NULL) +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.x = element_text(size = 8),
        axis.text.y = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        text = element_text(family = "arial"),
        plot.margin = margin(0, 0, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(-max_val, 0),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
                breaks = c(-0.00001, -0.00002, -0.00003)

    ) + 
  guides(fill = FALSE)

(total_plot <-  f_plot + p_plot)

ggdraw(total_plot) +
       draw_label("0", y = 0.085, x = 0.49, size = 8, fontfamily = "arial", color = "grey30") +
  draw_label("Proportion of global habitat area impacted", y = 0.03, x = 0.5, size = 8, fontfamily = "arial", color = "black") 


ggsave(here(this_dir, "/MS_plots/fig3_economic_barchart.png"), width = 178.9, height =100.92 , units = "mm", bg = "white")


```


SI Figure 5. Global average proportion of habitat area impacted in cells by raw material and diet. Related to Figure 3. 

SI Figure 6. Average proportion of habitat area impacted per 100% inclusion of each raw material. Values are normalised to reflect the habitat displacement that would occur if each material made up 100% of the formulation. This assumes a linear relationship with raw material inclusion. Related to Figure 3. 

```{r}

raw_materials <- c("Soybean", "Wheat", "Sunflower", "CropsNES", "Rapeseed", "Maize", "forage fish", "trimmings fish", "Pulses")

num_cores <- 9
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(raw_mat = raw_materials,
        .packages = c("qs", "terra", "tidyverse", "dplyr", "glue")) %dopar% {
for(allocation_type in c("economic", "mass", "ge")){
  for(fcr_type in fcr_types){
    for(diet_type in diet_types){
          
# diet_type = "plant-dominant"
# fcr_type = "regular"
# allocation_type = "economic"
# raw_mat = "Rapeseed"

          if(length(list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_mean_prop.tif"), full.names = TRUE)) > 0 ){
        
            files <- c(list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_mean_prop.tif"), full.names = TRUE), list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_nspp.tif"), full.names = TRUE))
            
        
        map_df <- rast(files) %>%
          as.data.frame(., xy = TRUE) %>%
          mutate(raw_material = raw_mat) %>%
          group_by(raw_material) %>%
          summarise(mean_prop_weighted = weighted.mean(mean_prop, nspp, na.rm = TRUE),
                    mean_prop = mean(mean_prop, na.rm = TRUE),
                    sum_prop = sum(mean_prop, na.rm = TRUE)) %>%
          mutate(allocation = allocation_type, 
                 diet = diet_type, 
                 fcr = fcr_type, 
                 raw_material = raw_mat) %>%
          ungroup()
        
        
        qsave(map_df, file.path(glue(impact_maps_dir, "/csvs/by_material_global_summary/mean_impacts_{allocation_type}_{diet_type}_{fcr_type}_{raw_mat}.qs")))
        
          }else{
            next()
          }
        
        }
      }
    }
  }

stopCluster(cl)

# now lets join all of those together into one dataframe and save 

formulations <- read.csv(here("analysis/int/ingredient_raw_mats_formulation.csv")) %>%
  clean_names() %>%
  group_by(raw_material) %>%
  summarise(fish_dominant = sum(fish_dominant, na.rm = TRUE)/100,
            plant_dominant = sum(plant_dominant, na.rm = TRUE)/100) %>%
  ungroup() %>%
  filter(raw_material != "") %>%
    pivot_longer(cols = c(fish_dominant, plant_dominant), names_to = "diet", values_to = "prop_inclusion") %>%
  mutate(diet = ifelse(diet == "plant_dominant", "Plant-dominant", "Fish-dominant")) %>%
  mutate(raw_material = ifelse(raw_material == "Trimmings", "Trimmings fish", raw_material))

csv_files <- list.files(file.path(impact_maps_dir, "csvs/by_material_global_summary"), pattern = ".qs", full.names = TRUE)

list_of_dfs <- lapply(csv_files, qread) 

combined_impact_df <- do.call(rbind, list_of_dfs) %>%
  mutate(raw_material = str_to_sentence(raw_material), 
         allocation = str_to_sentence(allocation),
         diet = str_to_sentence(diet),
         fcr = str_to_sentence(fcr))  %>%
 mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material)) %>%
  left_join(formulations) %>%
  mutate(impact_per_formulation = mean_prop/prop_inclusion)

material_stats_1 <- combined_impact_df %>%
    filter(allocation == "Economic") %>%
  filter(fcr == "Regular")

# Find the maximum absolute value
max_val <- max(abs(material_stats_1$mean_prop), na.rm = TRUE)

summary_df_p <- material_stats_1 %>% filter(diet == "Plant-dominant")

summary_df_p$raw_material <- with(summary_df_p, reorder(raw_material, mean_prop, FUN = sum))

# 4. Plot
p_plot <- ggplot(summary_df_p, aes(x = mean_prop, 
                                 y = raw_material, 
                                 fill = raw_material)) +
    geom_col() +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = NULL, 
         fill =  NULL) +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.y = element_blank(),
        axis.text.x = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8),
        text = element_text(family = "arial"),
        axis.ticks.y = element_blank(),
        plot.margin = margin(0, 10, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(0, max_val),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
        breaks = c(0.00005, 0.00010, 0.00015)
    ) +
    guides(fill = FALSE)

summary_df_f <- material_stats_1 %>% filter(diet == "Fish-dominant") %>%
  dplyr::select(raw_material, diet, mean_prop) %>%
  add_row(raw_material = "Rapeseed", diet = "Fish-dominant", mean_prop = 0) %>%
    add_row(raw_material = "Sunflower", diet = "Fish-dominant", mean_prop = 0) %>%
    add_row(raw_material = "Other crops", diet = "Fish-dominant", mean_prop = 0) %>%
  mutate(mean_prop = -1*mean_prop)


summary_df_f$raw_material <- factor(summary_df_f$raw_material, levels = levels(summary_df_p$raw_material))

# 4. Plot
f_plot <- ggplot(summary_df_f, aes(x = mean_prop, 
                                 y = raw_material, 
                                 fill = raw_material)) +
    geom_col() +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = "Raw material", 
         fill = NULL) +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.x = element_text(size = 8),
        axis.text.y = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        text = element_text(family = "arial"),
        plot.margin = margin(0, 0, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(-max_val, 0),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
        breaks = c(-0.00005, -0.00010, -0.00015)

    ) + 
  guides(fill = FALSE)

(total_plot <-  f_plot + p_plot)

ggdraw(total_plot) +
       draw_label("0", y = 0.085, x = 0.56, size = 8, fontfamily = "arial", color = "grey30") +
  draw_label("Average proportion of habitat area impacted", y = 0.03, x = 0.55, size = 8, fontfamily = "arial", color = "black") 

ggsave(here(this_dir, "SI_plots/SI_Fig_5_average_raw_mat_impact.png"), bg = "white", width = 178.9, height = 100.92 , units = "mm")

# Find the maximum absolute value
max_val <- max(abs(material_stats_1$impact_per_formulation), na.rm = TRUE)

summary_df_p <- material_stats_1 %>% filter(diet == "Plant-dominant")

summary_df_p$raw_material <- with(summary_df_p, reorder(raw_material, impact_per_formulation, FUN = sum))

# 4. Plot
p_plot <- ggplot(summary_df_p, aes(x = impact_per_formulation, 
                                 y = raw_material, 
                                 fill = raw_material)) +
    geom_col() +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = NULL, 
         fill =  NULL) +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.y = element_blank(),
        axis.text.x = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8),
        text = element_text(family = "arial"),
        axis.ticks.y = element_blank(),
        plot.margin = margin(0, 0, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(0, max_val),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
        breaks = c(0.0002, 0.0004, 0.0006, 0.0008)
    ) +
    guides(fill = FALSE)

summary_df_f <- material_stats_1 %>% filter(diet == "Fish-dominant") %>%
  dplyr::select(raw_material, diet, impact_per_formulation) %>%
  add_row(raw_material = "Rapeseed", diet = "Fish-dominant", impact_per_formulation = 0) %>%
    add_row(raw_material = "Sunflower", diet = "Fish-dominant", impact_per_formulation = 0) %>%
    add_row(raw_material = "Other crops", diet = "Fish-dominant", impact_per_formulation = 0) %>%
  mutate(impact_per_formulation = -1*impact_per_formulation)


summary_df_f$raw_material <- factor(summary_df_f$raw_material, levels = levels(summary_df_p$raw_material))

# 4. Plot
f_plot <- ggplot(summary_df_f, aes(x = impact_per_formulation, 
                                 y = raw_material, 
                                 fill = raw_material)) +
    geom_col() +
    facet_wrap(~diet, scales = "fixed") +  # No "scales = free_x" -- we keep same scale
    scale_fill_manual(values = pal_mats) +
    labs(x = NULL, 
         y = "Raw material", 
         fill = NULL) +
    theme_minimal() + 
    theme(
        panel.background = element_blank(),
        axis.text.x = element_text(size = 8),
        axis.text.y = element_text(size = 8),
        strip.text = element_text(size = 8),
        axis.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        text = element_text(family = "arial"),
        plot.margin = margin(0, 0, 15, 0)  # Remove any plot margins

    ) +
    scale_x_continuous(
        limits = c(-max_val, 0),  # Fix limits symmetrically
        expand = c(0, 0),
        labels = abs,  # Label all ticks as positive
        breaks = c(-0.0002, -0.0004, -0.0006, -0.0008)

    ) + 
  guides(fill = FALSE)

(total_plot <-  f_plot + p_plot)

ggdraw(total_plot) +
       draw_label("0", y = 0.085, x = 0.57, size = 8, fontfamily = "arial", color = "grey30") +
  draw_label("Average proportional habitat area impact per 100% material inclusion", y = 0.03, x = 0.6, size = 8, fontfamily = "arial", color = "black") 

ggsave(here(this_dir, "SI_plots/SI_Fig_6_material_normalized.png"), bg = "white", width = 178.9, height = 100.92 , units = "mm")

```

SI Figure 7. Comparison of raw material contribution to diet formulation vs. total impacts. Related to T
Table 1 and Figure 3. 

```{r}
codes <- read.csv(here("prep/02_feed/data/ingredient_spam_fao_codes.csv")) %>%
  dplyr::select(raw_name = source_ingredient, raw_material = raw.material)

diets_raw_mat <- read.csv(here("prep/02_feed/data/aquaculture_diet_composition.csv")) %>%
  distinct(raw_name, prop_diet, diet) %>%
  left_join(codes) %>%
  mutate(raw_material = case_when(
    str_detect(raw_name, "cut offs") ~ "Trimmings fish", 
    str_detect(raw_name, "pea flour|guar|pea pro|faba") ~ "Pulses",
    str_detect(raw_name, "forage fish") ~ "Forage fish",
    str_detect(raw_name, "coconut|linseed") ~ "Other crops",
    TRUE ~ str_to_title(raw_material)
    
  )) %>%
  group_by(raw_material, diet) %>%
  summarise(total_diet_contribution = sum(prop_diet, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(!is.na(raw_material)) %>%
  mutate(diet = str_to_sentence(diet))

summary_df_barcharts <- all_overlap_df %>%
  group_by(diet, fcr, raw_material, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_hab_area) %>%
  mutate(prop_impact = impact_total/total_hab_area)


summary_df_stats <- summary_df_barcharts %>%
   group_by(diet, fcr, raw_material, allocation) %>%
  summarise(prop_total_impact = sum(prop_impact),
            total_km2 = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr, allocation) %>% 
  mutate(total_impact = sum(prop_total_impact),
         total_km2_scenario = sum(total_km2, na.rm = TRUE)) %>%
  mutate(prop_attributed = prop_total_impact/total_impact,
         prop_total_km2 = total_km2/total_km2_scenario) %>%
  left_join(diets_raw_mat) %>%
  mutate(delta = prop_attributed - total_diet_contribution) %>%
  mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material)) %>%
  mutate(km2_per_formulation = total_km2/total_diet_contribution) %>%
  filter(prop_attributed > 0 )
  

## test out a plot 
pal_mats <- c("Soybean" = "#328C41",
              "Trimmings fish" = "#8FBAED",
              "Sunflower" = "#4277A9",
              "Wheat" = "#FFE96D",
              "Rapeseed" = "#AC9900",
              "Pulses" = "#79E0DA",
              "Forage fish" = "#3DB1AB",
              "Maize" = "#0A6B66",
              "Other crops" = "#BC94C6")

ggplot(summary_df_stats, aes(x = prop_total_km2, y = total_diet_contribution, color = raw_material)) +
  geom_point(size = 3) +
  facet_wrap(~ diet) +
  labs(color = "",
       x = "Impact Contribution",
       y = "Diet Contribution") +
  theme_bw() +
  theme(legend.position = "bottom") +
  geom_abline() +
  scale_color_manual(values = pal_mats)

ggsave(here("analysis/SI_plots/SI_Fig_7_efficiency.png"), width = 178.9, height =100.92 , units = "mm", bg = "white")

```


EXTRA PLOTS: 

**Look at per country land/eez impacts**
 - grab by raw material rasters 
 - currently we are taking the average of an average... is there a way to fix this without great computational cost? I think taking a weighted average by number of impacted species is ok, thanks ChatGPT: 
 'Weighting by the number of species impacted helps mitigate any distortion from taking an average of an average. It ensures that cells with more species (and thus potentially higher impact) contribute more to the overall country-level impact.
If the weights are used appropriately in the second step, this avoids the common issue of diluting the information when averaging an average. The weighted average ensures that the relative importance of each cell is preserved in the final country-level estimate."
 

```{r}
allocations <- c("economic")
fcr_types <- c("regular")
diet_types <- c("plant-dominant", "fish-dominant")
spp_types <- all_overlap %>% 
  filter(spp_type != "polychaetes") %>% 
  pull(spp_type) %>% unique()

ingredient_types <- all_overlap %>% 
  pull(ingredient) %>% unique()

raw_materials <- c("Soybean", "Wheat", "Sunflower", "CropsNES", "Rapeseed", "Maize", "forage fish", "trimmings fish", "Pulses")

land_eez_rgns <- read.csv(file.path(biodiv_dir, "spatial/rgns_xy_all.csv")) %>% 
  dplyr::select(x, y, rgn_id)

land_eez_rgns_ids <- read.csv(here("data/spatial/output/region_ids.csv"))
setdiff(land_eez_rgns_ids$rgn_id, land_eez_rgns$rgn_id) # 0 good

num_cores <- 9
cl <- makeCluster(num_cores)
registerDoParallel(cl)


foreach(raw_mat = raw_materials,
        .packages = c("qs", "terra", "tidyverse", "dplyr", "glue")) %dopar% {
for(allocation_type in c("economic", "mass", "ge")){
  for(fcr_type in fcr_types){
    for(diet_type in diet_types){
          
# diet_type = "plant-dominant"
# fcr_type = "regular"
# allocation_type = "economic"
# raw_mat = "Rapeseed"

          if(length(list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_mean_prop.tif"), full.names = TRUE)) > 0 ){
        
            files <- c(list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_mean_prop.tif"), full.names = TRUE), list.files(file.path(glue(impact_maps_dir, "/impact_maps_across_taxon_by_material/{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_{raw_mat}_nspp.tif"), full.names = TRUE))
            
        
        map_df <- rast(files) %>%
          as.data.frame(., xy = TRUE) %>%
          left_join(land_eez_rgns) %>%
          group_by(rgn_id) %>%
          summarise(mean_prop_weighted = weighted.mean(mean_prop, nspp, na.rm = TRUE),
                    mean_prop = mean(mean_prop, na.rm = TRUE),
                    sum_prop = sum(mean_prop, na.rm = TRUE)) %>%
          mutate(allocation = allocation_type, 
                 diet = diet_type, 
                 fcr = fcr_type, 
                 raw_material = raw_mat) %>%
          ungroup()
        
        
        qsave(map_df, file.path(glue(impact_maps_dir, "/csvs/by_material_country_summary/mean_impacts_{allocation_type}_{diet_type}_{fcr_type}_{raw_mat}.qs")))
        
          }else{
            next()
          }
        
        }
      }
    }
  }

stopCluster(cl)


rgn_ids <- read.csv(here("data/spatial/output/region_ids.csv"))

# now lets join all of those together into one dataframe and save 

csv_files <- list.files(file.path(impact_maps_dir, "csvs/by_material_country_summary"), pattern = ".qs", full.names = TRUE)

list_of_dfs <- lapply(csv_files, qread) 

combined_impact_df <- do.call(rbind, list_of_dfs) %>%
  left_join(rgn_ids) %>%
  mutate(raw_material = str_to_sentence(raw_material), 
         allocation = str_to_sentence(allocation),
         diet = str_to_sentence(diet),
         fcr = str_to_sentence(fcr))  %>%
 mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material))

## do the same thing but make a plot with raw material as the fill value

summary_df <- combined_impact_df %>%
  dplyr::select(diet, iso3c, rgn_id, fcr, allocation, raw_material, mean_prop, mean_prop_weighted) %>%
  filter(allocation == "Economic", fcr == "Regular") 
  
  top_25_impacts <- summary_df %>% 
   group_by(iso3c, diet, fcr) %>%
  summarise(total_impact = sum(mean_prop_weighted, na.rm = TRUE)) %>%
    ungroup() %>%
    group_by(diet, fcr) %>% 
    arrange(-total_impact) %>%
    mutate(rank = row_number()) %>%
    filter(rank <= 25) %>%
    ungroup() %>%
    dplyr::select(-total_impact) 
  
  plot_df <- summary_df %>%
    group_by(iso3c, diet, fcr) %>%
    filter(iso3c %in% c(top_25_impacts$iso3c)) %>%
  group_by(iso3c, diet, fcr) %>%
  mutate(total_impact = sum(mean_prop_weighted, na.rm = TRUE)) %>%
    left_join(top_25_impacts) %>%
    filter(!is.na(rank))
  
legend_plot <-  get_legend(ggplot(summary_df, aes(x = iso3c, y = mean_prop_weighted, fill = raw_material)) +
    geom_bar(stat = "identity") +
  facet_wrap(~diet,  as.table = TRUE, drop = TRUE, shrink = TRUE) +
 scale_fill_manual(values = pal_mats) + 
  theme_minimal() + 
  coord_flip()  + 
facet_grid(fcr~diet, scales = "free") +
    theme(axis.text.x = element_text(angle = -90, hjust = 1),
          axis.text.y = element_text( size = 5), 
          legend.title = element_blank()))

  

panel_1_df <- plot_df %>% 
  filter(diet == "Fish-dominant")

panel_1_df$iso3c <- with(panel_1_df, reorder(iso3c, total_impact, FUN = function(x) mean(x)))


panel_1 <- ggplot(panel_1_df, aes(x = iso3c, y = mean_prop_weighted, fill = raw_material)) +
    geom_bar(stat = "identity") +
 scale_fill_manual(values = pal_mats) + 
  theme_minimal() + 
  coord_flip()  + 
facet_grid(fcr~diet, scales = "free") +
    theme(axis.text.x = element_text(angle = -90, hjust = 1, size = 8),
          axis.text.y = element_text( size = 8), 
          legend.position = "none", 
          axis.title.x = element_blank(), 
          strip.text.y = element_blank()) + # Adjust the 'angle' value as needed
  labs(x = "")

panel_2_df <- plot_df %>% 
  filter(diet == "Plant-dominant")

panel_2_df$iso3c <- with(panel_2_df, reorder(iso3c, total_impact, FUN = function(x) mean(x)))


panel_2 <- ggplot(panel_2_df, aes(x = iso3c, y = mean_prop_weighted, fill = raw_material)) +
    geom_bar(stat = "identity") +
 scale_fill_manual(values = pal_mats) + 
  theme_minimal() + 
  coord_flip()  + 
facet_wrap(~diet) +
    theme(axis.text.x = element_text(angle = -90, hjust = 1),
          axis.text.y = element_text( size = 8), 
          legend.position = "none", 
          axis.title.x = element_blank(), 
          axis.title.y = element_blank()) 


total_panel <- panel_1 + panel_2

ggdraw() + 
  draw_plot(total_panel) +
  draw_label('Average proportion of habitat area impacted', x = 0.35, y = .034, hjust = 0, vjust = 1, 
             size = 11, fontfamily = 'Helvetica') +
    theme(plot.margin = margin(0, 3, 0, 0, "cm")) + # Adjust the right margin as needed
    draw_plot(legend_plot,  x = 1.02, y = 0.5, height = .10, width = .15)

ggsave(here(this_dir, "SI_plots/extra_plots/economic_country_props.png"), bg = "white", width = 178.9, height =100.92 , units = "mm")

country_stats_1 <- combined_impact_df %>%
  filter(allocation == "Economic") %>%
  group_by(diet, fcr, iso3c) %>%
  summarise(sum_prop_mean = sum(mean_prop_weighted, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr) %>%
  mutate(total_prop_mean = sum(sum_prop_mean, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impacts = sum_prop_mean/total_prop_mean)
```



**Plot showing diet percentages**
 - We didn't end up using this. 
 
```{r}
diet_comps <- read.csv(here("prep/02_feed/data/aquaculture_diet_composition.csv")) %>%
  distinct(raw_name, prop_diet, diet) %>%
  filter(prop_diet > 0)

# check sums 
diet_comps %>% filter(diet == "fish-dominant") %>% pull(prop_diet) %>% sum()
diet_comps %>% filter(diet == "plant-dominant") %>% pull(prop_diet) %>% sum()


ingredient_levels <- c( "Fish meal, forage fish",  "Fish oil, forage fish", "Fish meal, trimmings", "Fish oil, trimmings", "Soybean meal", "Corn gluten meal" , "Faba beans" , "Wheat gluten" ,"Wheat" , "Soy protein concentrate",  "Guar meal",  "Pea protein concentrate", "Sunflower meal", "Canola oil", "Linseed oil", "Soy oil", "Pea flour", "Coconut oil", "Micro ingredients")


both_diets <- diet_comps %>% 
  mutate(raw_name = str_replace_all(raw_name, "cut offs", "trimmings")) %>%
      mutate(hjust = if_else(condition = diet == "fish-dominant", true = 1, false = 0)) |> 
    mutate(label_x_pos = if_else(diet =="fish-dominant", true =0.5, false = 2.5)) |> 
  mutate(diet = if_else(condition = diet == "fish-dominant", true = "Fish-\ndominant", false = "Plant-\ndominant")) %>% 
   mutate(diet = factor(diet, levels = rev(c("Plant-\ndominant", "Fish-\ndominant")))) %>%
  filter(raw_name != "other") %>%
  mutate(ingredients = str_to_sentence(raw_name)) %>% 
  mutate(ingredients = factor(ingredients, levels = rev(ingredient_levels))) 


ggplot(data = both_diets,
       aes(x = diet , y = prop_diet, fill = ingredients))+
  geom_col()+
  scale_fill_manual(values =  c("darkgoldenrod4", rev(colorRampPalette(brewer.pal(n=9, name = "Greens"))(14)),  brewer.pal(n=5, name = "Blues")[c(2:5)]))+
  theme_pubr()+
  theme(legend.title = element_blank(),
        legend.text = element_text(size=7),
        text = element_text(size=8),
        legend.position = "right",
        #legend.box.spacing = unit(-0.1, "cm"),
        panel.grid = element_blank(),
        legend.key.size = unit(0.25, "cm")
        )+
  labs(x = "Feed scenario", y = "Proportion")+
  guides(fill = guide_legend(byrow=TRUE, reverse = FALSE, ncol = 1))

ggsave(filename = here(this_dir, "SI_plots/extra_plots/feed_composition.jpg"), dpi = 600, width = 10, height = 9, units = "cm")


```
