==> analysis/04_SI_IUCN.Rmd <==
---
title: "IUCN category plots"
output: html_document
editor_options: 
  chunk_output_type: console
---
##  Summary

In this script we create supplementary figure 4. In this figure, we split impacts by taxon and species IUCN risk category:

> Number of species within each impact quantile and taxon, split by current IUCN status and feed formulation scenario. Percentage wheels show the percentage of species from each assessed taxonomic group that are impacted, and thus, shown in the plot. It is well established that the IUCN has major data deficiencies, particularly in marine species 1, and we find that approximately 33% of the species we assessed lack adequate evaluation (data deficient or not evaluated), the majority of which (~83%) are marine (Figure 4). Least concern (LC) species predominate in both dietary scenarios and across all impact categories for all taxa except molluscs, arthropods, and other marine species, which are majority data deficient or not evaluated. The amphibian taxon has a high percentage of species falling into categories that are of a high threat level (i.e., above that of Least Concern). For example, under the plant-dominant diet, 43% of amphibian species represented in the 75-94th impact quantile are at-risk of extinction. However, most often critically endangered and endangered species are in the 25-49th impact category for both fish- and plant-dominant dietary scenarios. Evaluating these impacts against IUCN categories can guide conservation efforts by identifying at-risk taxa which could benefit most from adjusting production areas or raw material types used in feeds. 

## Setup

Load libraries

```{r setup, include=FALSE}
library(tidyverse)
library(here)
library(strex)
library(janitor)
library(tools)
library(patchwork)
library(cowplot)
library(glue)
library(RColorBrewer)
library(ggpubr)
library(smoothr)
library(rredlist)
library(ggpattern)
library(ggpp)
library(scales)

source(here("src/directories.R"))

source(here("src/spatial.R"))
source(here("src/fxns.R"))

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

impact_maps_dir <- file.path(biodiv_dir, "output")

```

```{r}
# Create a named vector with colors for each spp_type
color_scale <- readRDS(here("prep/03_prep_spp_habitats/int/color_scale.RDS"))

pal_mats <- c("Soybean" = "#328C41",
              "Trimmings fish" = "#8FBAED",
              "Sunflower" = "#4277A9",
              "Wheat" = "#FFE96D",
              "Rapeseed" = "#AC9900",
              "Pulses" = "#79E0DA",
              "Forage fish" = "#3DB1AB",
              "Maize" = "#0A6B66",
              "Cropsnes" = "#BC94C6")

theme_ohara <- function(base_size = 9) {
  theme(axis.ticks = element_blank(),
        text             = element_text(family = 'Helvetica',
                                        color = 'gray30', size = base_size),
        plot.title       = element_text(size = rel(1.25), hjust = 0, face = 'bold'),
        panel.background = element_blank(),
        legend.position  = 'right',
        panel.border     = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(colour = 'grey90', size = .25),
        legend.key       = element_rect(colour = NA, fill = NA),
        axis.line        = element_blank())
}

```

Read in marine and terrestrial dfs and combine

```{r}
marine_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds")) %>%
     filter(sensitivity_scenario == "original") %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)


terrestrial_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_terrestrial_impact.rds")) %>% 
   filter(sensitivity_scenario == "original") %>%
      dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

all_overlap <- rbind(marine_overlap, terrestrial_overlap) %>%
    mutate(taxon = str_to_sentence(spp_type)) %>%
  mutate(taxon = case_when(
    taxon == "Terrestrial mammal" ~ "Terrestrial mammals",
    taxon == "Bird" ~ "Birds",
    taxon == "Fish" ~ "Finfish", 
    taxon == "Marine mammal" ~ "Marine mammals",
    taxon == "Marine plant" ~ "Marine plants",
    TRUE ~ taxon)) %>%
  mutate(sciname = tolower(sciname))

length(unique(all_overlap$sciname)) # 54622

spp_habs <- all_overlap %>%
  distinct(sciname, taxon, total_hab_area) %>%
  group_by(sciname, taxon) %>%
  summarise(total_hab_area = max(total_hab_area)) %>%
  ungroup()

length(unique(spp_habs$sciname)) # 54622


```

Get IUCN categories for species

```{r}

spp_names_df <- all_overlap %>%
  distinct(sciname = tolower(sciname), taxon)


out <- rl_sp(all = TRUE, key = api_key) 

length(out) 

all_df <- do.call(rbind, lapply(out, "[[", "result")) %>%
  mutate(sciname = tolower(scientific_name)) %>%
  dplyr::select(sciname, category)

spp_cats <- spp_names_df %>%
  left_join(all_df) %>%
  distinct(sciname, taxon, category) %>%
  mutate(category = case_when(
    is.na(category) ~ "NE",
    category == "LR/cd" ~ "NT",
    category == "LR/lc" ~ "LC",
     category == "LR/nt" ~ "NT",
    TRUE ~ category
  )) %>%
    distinct(sciname, taxon, category) %>%
  mutate(rank = case_when(
    category %in%  c("NE") ~ 0,
    category == "DD" ~ 1,
    category == "LC" ~ 2,
    category == "NT" ~ 3,
    category == "VU" ~ 4,
    category == "EN" ~ 5, 
    category == "CR" ~ 6,
    category == "EW" ~ 7,
    category == "EX" ~ 8
  )) %>%
  group_by(sciname, taxon) %>%
  summarise(rank = max(rank)) %>% ## we're just going to take the worst category if it has multiple for some reason...
  ungroup() %>%
    mutate(category = case_when(
   rank ==  0 ~ "NE",
    rank == 1 ~ "DD",
    rank == 2 ~ "LC",
    rank == 3 ~ "NT",
    rank == 4 ~ "VU",
    rank == 5 ~ "EN", 
    rank == 6 ~ "CR",
    rank == 7 ~ "EW",
    rank == 8 ~ "EX"
  )) %>%
   mutate(id = row_number()) %>%
  dplyr::select(-rank)

write.csv(spp_cats, here("prep/03_prep_spp_habitats/int/spp_iucn_cats.csv"), row.names = FALSE)
 
unique(spp_cats$category) #  [1] "NE" "LC" "DD" "VU" "NT" "EN" "CR" "EX" "EW"
  ## DD = data deficient
  ## LC = least concern
  ## VU = vulnerable
  ## NT = near threatened
  ## EN = endangered
  ## CR = critically endangered
  ## LR/cd = near threatened
  ## LR/nt = near threatened
  ## EX = Extinct 
  ## LR/lc = least concern
  ## EW = extinct in the wild
  ## NE = not evaluated




```

Match IUCN categores to our overlap data and match impacts to quantile ranges

```{r}

spp_cats <- read.csv(here("prep/03_prep_spp_habitats/int/spp_iucn_cats.csv")) %>%
  dplyr::select(-id)

all_overlap_summary <- all_overlap %>%
  group_by(sciname, taxon, diet, fcr_type, allocation) %>%
  summarise(impact_km2 = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(spp_habs) %>%
    left_join(spp_cats) %>%
  mutate(prop_impact = impact_km2/total_hab_area) %>%
  group_by(allocation, fcr_type) %>%
  mutate(
    quantile_25 = quantile(prop_impact, probs = 0.25),
    quantile_50 = quantile(prop_impact, probs = 0.50),
    quantile_75 = quantile(prop_impact, probs = 0.75),
    quantile_95 = quantile(prop_impact, probs = 0.95),
    quantile_99 = quantile(prop_impact, probs = 0.99)
  ) %>%
  ungroup() %>%
  mutate(impact_category = case_when(
    prop_impact < quantile_25 ~ "<25",
    prop_impact >= quantile_25 & prop_impact < quantile_50 ~ "25-49",
    prop_impact >= quantile_50 & prop_impact < quantile_75 ~ "50-74",
   prop_impact >= quantile_75 & prop_impact < quantile_95 ~ "75-94",
   prop_impact >= quantile_95 ~ "≥95",
   TRUE ~ NA
  )) %>%
  mutate(iucn_cat = case_when(
    category == "LC" ~ "Least Concern",
    category == "NT" ~ "Near Threatened",
    category == "DD" ~ "Data Deficient",
    category == "NE" ~ "Not Evaluated",
    category == "VU" ~ "Vulnerable", 
    category == "EN" ~ "Endangered",
    category == "CR" ~ "Critically Endangered",
    category == "EX" ~ "Extinct", 
    category == "EW" ~ "Extinct In The Wild"
  )) %>%
  filter(!(category %in% c("EX", "EW"))) %>%
  mutate(taxon_plot = case_when(
    taxon %in% c("Cephalopods", "Cnidaria", "Corals", "Echinoderms", "Marine mammals", "Marine plants", "Sponges", "Elasmobranchs") ~ "Other marine species", 
    TRUE ~ taxon
  )) # now we need to group some of the marine species into "other marine species" category

```


Make plots of the % of species in each taxon impacted (we only want to show impacted spp in the main plot)
 - We will add these circle plots to the main plot in post processing, since I can't figure out how to do it in R. 

```{r}

nspp_df <- all_overlap_summary %>%
    filter(allocation == "economic",
         fcr_type == "regular", 
         diet == "plant-dominant" # just filter for one diet as same amount assessed regardless
         ) %>%
  group_by(taxon_plot) %>%
  summarise(n_spp_assessed  = n_distinct(sciname)) %>%
  ungroup()

 sum(nspp_df$n_spp_assessed) # 54615 - ok cool
 
 quantile_df <- all_overlap_summary %>%
  filter(allocation == "economic",
         fcr_type == "regular"
         ) %>%
  filter(impact_km2 > 0) %>% 
   distinct(diet, fcr_type, quantile_25, quantile_50, quantile_75, quantile_95, quantile_99)

plot_df <- all_overlap_summary %>%
  filter(allocation == "economic",
         fcr_type == "regular"
         ) %>%
  filter(impact_km2 > 0) %>% 
  group_by(diet, category, iucn_cat, taxon_plot, impact_category) %>%
  summarise(n_spp = n_distinct(sciname)) %>%
  ungroup() %>%
  left_join(nspp_df) %>%
  group_by(diet, taxon_plot) %>%
  mutate(total_spp_impacted = sum(n_spp)) %>%
  ungroup()

nspp_props <- plot_df %>% 
  distinct(taxon_plot, diet, n_spp_assessed, total_spp_impacted) %>%
  mutate(prop_impacted = round(total_spp_impacted/n_spp_assessed, 3)) %>%
  dplyr::select(-n_spp_assessed, -total_spp_impacted) %>%
      mutate(percent_imp = percent(prop_impacted))

# max_prop <- data.frame(taxon_plot = "test taxa", diet = "plant-dominant", prop_impacted = 1, percent_imp = "100%")

diet_cols = c("fish-dominant" = "#474747", "plant-dominant" = "#474747") # fd is dark grey, pd is grey?

for(taxa in unique(plot_df$taxon_plot)){

#  taxa = "Amphibians"
nspp_props_loop <- nspp_props %>% 
  filter(taxon_plot == taxa) 
  
p <- ggplot(nspp_props_loop,
         aes(x = 1,
             y = prop_impacted,
             fill = diet, 
             label = percent_imp)) +
  geom_col() +
    geom_col_pattern(data = nspp_props_loop %>% filter(diet == "plant-dominant"),
                                             aes(pattern_fill = diet),
                     pattern_alpha = 0.3,
             # pattern_size = 0.5, pattern_spacing = 0.025,
             pattern = "stripe", pattern_color = "white", pattern_fill = "white") +
    coord_polar(theta = "y", direction = -1, clip = "off") +
      theme_void() +
    scale_fill_manual(values = diet_cols, breaks = names(diet_cols)) +
   theme(
        # strip.text = element_blank(), 
         strip.background = element_blank(),
         panel.spacing.x = unit(-25, "mm"),
         panel.spacing.y = unit(-15, "mm")
         ) + 
   facet_grid(diet ~ taxon_plot) +
   geom_text(x = -2, y = -2, size = 9) +
guides(fill = "none") +
  scale_y_continuous(expand = c(0,0), limits = c(0,1)) +
  scale_x_continuous(expand = c(0,0), limits = c(-2, 2))


ggsave(plot = p, glue(here("analysis/SI_plots/taxon_props/{taxa}_props_vert.png")), units = "mm", width = 180, height = 100)

}

# W : 25.204
# H : 12.628

# W: 9.634
# H: 19.735
```


Now make IUCN plot, only with >0 impact species

```{r}

plot_df_new <- plot_df

x_axis_order <- c("<25", "25-49", "50-74", "75-94", "≥95")
plot_df_new$impact_category <- factor(plot_df_new$impact_category, levels = x_axis_order)

fill_order <- c("Least Concern", "Near Threatened", "Data Deficient", "Not Evaluated", "Vulnerable", "Endangered", "Critically Endangered")

plot_df_new$iucn_cat <- factor(plot_df_new$iucn_cat, levels = fill_order)

plot_df_new$taxon_plot <- factor(plot_df_new$taxon_plot, levels = c("Birds", "Finfish", "Reptiles", "Molluscs", "Amphibians", "Terrestrial mammals", "Arthropods", "Other marine species"))

iucn_cat_pal <- c(
  "Least Concern" = "#54B050",
  "Near Threatened" = "#C9CE3A", 
  "Data Deficient" = "#C5C5C5", 
  "Not Evaluated" = "darkgray",
  "Vulnerable" = "#FED339", 
  "Endangered" = "#FD643E", 
  "Critically Endangered" = "#D40A1E"
)

iucn_plot <- ggplot() +
  geom_col(plot_df_new %>% filter(diet == "fish-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat), 
               width = 0.4, position_stacknudge(x = -0.21)) +
    geom_col_pattern(plot_df_new %>% filter(diet == "plant-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat, pattern_fill = diet), 
               width = 0.4, position = position_stacknudge(x = 0.21), pattern_alpha = 0.3,
             pattern_size = 0.5, pattern_spacing = 0.025, pattern = "stripe") + 
  facet_wrap(~taxon_plot, nrow = 2, scales = "free_x") +
  labs(
       x = "Impact Quantile",
       y = "Number of Species") +
  theme_ohara() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 7, family = "Arial"),
        axis.text.y = element_text(size = 7, family = "Arial", colour = "black"),
        axis.title = element_text(size = 7, family = "Arial", colour = "black"),
        strip.text = element_text(size = 7, family = "Arial", colour = "black"), 
        text = element_text(family = "Arial", size = 7, colour = "black"),
        legend.position  = "none",
        plot.margin = margin(t = 0.25, r = 0.25, b = 3, l = 0.25, unit = "cm")) +
  scale_fill_manual(values = iucn_cat_pal)

legend_iucn <- get_legend( ggplot() +
  geom_col(plot_df_new %>% filter(diet == "fish-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat)) +
              theme(legend.position = "bottom",
                    legend.spacing.x = unit(0.28, 'cm')) +
    labs(fill = "Current IUCN Category") + 
    guides(fill = guide_legend(title.position = "bottom", title.hjust = 0.5)) +
      scale_fill_manual(values = iucn_cat_pal) + 
    theme(legend.text = element_text(size = 7, family = "Arial", colour = "black"), 
          legend.title = element_text(size = 7, colour = "black"), 
          legend.key.size = unit(5, "mm")))  

legend_pattern <-  get_legend(ggplot() + geom_col_pattern(plot_df_new,
             mapping = aes(x = impact_category, y = n_spp, pattern = diet), 
               width = 0.4, position = position_stacknudge(x = 0.21), pattern_alpha = 0.3,
             pattern_size = 0.5, pattern_spacing = 0.025, pattern_key_scale_factor = 0.5, pattern_color = "white") +
               theme(legend.position = "bottom",
                     legend.spacing.x = unit(0.28, 'cm'),
                     legend.text = element_text(size = 7, family = "Arial", colour = "black"),
                     legend.title = element_text(size = 7, colour = "black"), 
                     legend.key.size = unit(5, "mm")) + 
                 labs(pattern = "Diet") + 
                guides(pattern = guide_legend(title.position = "bottom", title.hjust = 0.5, nrow = 1)) +
               scale_pattern_manual(values=c('fish-dominant' = 'none', 'plant-dominant' = 'stripe'), 
                                    labels = c("Fish-dominant", "Plant-dominant")))


p <- ggdraw() + draw_plot(iucn_plot) + draw_plot(legend_iucn, x = 0.1, y = 0.005, width = 0.5, height = 0.15) + draw_plot(legend_pattern, x = 0.6, y = 0.005, width = 0.5, height = 0.118)

ggsave(here("analysis/SI_plots/SI_Fig_8_iucn_plot_no_props.png"), units = "mm", width = 180, height = 180)

## ok, now we will add the proportions created above in post processing, using inkscape; https://inkscape.org/

```

Save plot with 0 impact species 

```{r}

plot_df <- all_overlap_summary %>%
  filter(allocation == "economic",
         fcr_type == "regular"
         ) %>%
  group_by(diet, category, iucn_cat, taxon_plot, impact_category) %>%
  summarise(n_spp = n_distinct(sciname)) %>%
  ungroup()

plot_df_new <- plot_df

x_axis_order <- c("<25", "25-49", "50-74", "75-94", "≥95")
plot_df_new$impact_category <- factor(plot_df_new$impact_category, levels = x_axis_order)

fill_order <- c("Least Concern", "Near Threatened", "Data Deficient", "Not Evaluated", "Vulnerable", "Endangered", "Critically Endangered")
plot_df_new$iucn_cat <- factor(plot_df_new$iucn_cat, levels = fill_order)

plot_df_new$taxon_plot <- factor(plot_df_new$taxon_plot, levels = c("Birds", "Finfish", "Reptiles", "Molluscs", "Amphibians", "Terrestrial mammals", "Arthropods", "Other marine species"))

iucn_cat_pal <- c(
  "Least Concern" = "#54B050",
  "Near Threatened" = "#C9CE3A", 
  "Data Deficient" = "#C5C5C5", 
  "Not Evaluated" = "darkgray",
  "Vulnerable" = "#FED339", 
  "Endangered" = "#FD643E", 
  "Critically Endangered" = "#D40A1E"
)

iucn_plot <- ggplot() +
  geom_col(plot_df_new %>% filter(diet == "fish-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat), 
               width = 0.4, position_stacknudge(x = -0.21)) +
    geom_col_pattern(plot_df_new %>% filter(diet == "plant-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat, pattern_fill = diet), 
               width = 0.4, position = position_stacknudge(x = 0.21), pattern_alpha = 0.3,
             pattern_size = 0.5, pattern_spacing = 0.025, pattern = "stripe") + 
  facet_wrap(~taxon_plot, nrow = 2, scales = "free_x") +
  labs(
       x = "Impact Quantile",
       y = "Number of Species") +
  theme_ohara() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 7, family = "Arial"),
        axis.text.y = element_text(size = 7, family = "Arial", colour = "black"),
        axis.title = element_text(size = 7, family = "Arial", colour = "black"),
        strip.text = element_text(size = 7, family = "Arial", colour = "black"), 
        text = element_text(family = "Arial", size = 7, colour = "black"),
        legend.position  = "none",
        plot.margin = margin(t = 0.25, r = 0.25, b = 3, l = 0.25, unit = "cm")) +
  scale_fill_manual(values = iucn_cat_pal)

legend_iucn <- get_legend( ggplot() +
  geom_col(plot_df_new %>% filter(diet == "fish-dominant"),
             mapping = aes(x = impact_category, y = n_spp, fill = iucn_cat)) +
              theme(legend.position = "bottom",
                    legend.spacing.x = unit(0.28, 'cm')) +
    labs(fill = "Current IUCN Category") + 
    guides(fill = guide_legend(title.position = "bottom", title.hjust = 0.5)) +
      scale_fill_manual(values = iucn_cat_pal) + 
    theme(legend.text = element_text(size = 7, family = "Arial", colour = "black"), 
          legend.title = element_text(size = 7, colour = "black"), 
          legend.key.size = unit(5, "mm")))  

legend_pattern <-  get_legend(ggplot() + geom_col_pattern(plot_df_new,
             mapping = aes(x = impact_category, y = n_spp, pattern = diet), 
               width = 0.4, position = position_stacknudge(x = 0.21), pattern_alpha = 0.3,
             pattern_size = 0.5, pattern_spacing = 0.025, pattern_key_scale_factor = 0.5, pattern_color = "white") +
               theme(legend.position = "bottom",
                     legend.spacing.x = unit(0.28, 'cm'),
                     legend.text = element_text(size = 7, family = "Arial", colour = "black"),
                     legend.title = element_text(size = 7, colour = "black"), 
                     legend.key.size = unit(5, "mm")) + 
                 labs(pattern = "Diet") + 
                guides(pattern = guide_legend(title.position = "bottom", title.hjust = 0.5, nrow = 1)) +
               scale_pattern_manual(values=c('fish-dominant' = 'none', 'plant-dominant' = 'stripe'), 
                                    labels = c("Fish-dominant", "Plant-dominant")))


p <- ggdraw() + draw_plot(iucn_plot) + draw_plot(legend_iucn, x = 0.1, y = 0.005, width = 0.5, height = 0.15) + draw_plot(legend_pattern, x = 0.6, y = 0.005, width = 0.5, height = 0.118)

ggsave(here("analysis/SI_plots/extra_plots/SI_iucn_plot_zeros.png"), units = "mm", width = 180, height = 180)


```

```{r}
test <- plot_df %>% group_by(diet, category) %>% summarise(nspp = sum(n_spp)) %>% ungroup()

test2 <- plot_df %>% group_by(diet, category, taxon_plot) %>% summarise(nspp = sum(n_spp)) %>% ungroup() %>% filter(diet == "plant-dominant") %>% filter(category %in% c("NE", "DD"))

sum(test2$nspp) # 17763
17640/54628 # 0.3229113

test2 %>%
  filter(!(taxon_plot %in% c("Amphibians", "Birds", "Terrestrial mammals", "Reptiles"))) %>%
  pull(nspp) %>%
  sum()
14488/17763 # 0.815628

test3 <- plot_df  %>%
  group_by(taxon_plot, diet, impact_category) %>%
  mutate(total_spp = sum(n_spp)) %>%
  ungroup() %>%
  mutate(prop = n_spp/total_spp) %>% 
  mutate(category_2 = case_when(
    category == "LC" ~ "LC",
    category %in% c("CR", "EN", "CR", "NT", "VU") ~ "Not LC",
    TRUE ~ category
  )) %>%
    group_by(taxon_plot, diet, impact_category, category_2) %>%
    summarise(prop = sum(prop,na.rm = TRUE)) %>%
  ungroup() 


test <- plot_df %>% group_by(diet, category, impact_category) %>% summarise(nspp = sum(n_spp)) %>% filter(impact_category %in% c("≥95"))

test <- plot_df %>% group_by(diet, impact_category, category) %>% summarise(nspp = sum(n_spp)) %>% ungroup() %>% filter(category %in% c("EN", "CR"))
```
