==> analysis/05_publication_stats.Rmd <==
---
title: "Summary stats for manuscript"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary
This script contains code and data to fill out the various statistics requested for the main manuscript. Above each chunk is either an excerpt from the paper, or a specific request for a statistic.


## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(strex)
library(sf)
library(glue)
library(qs)
library(mapview)

source(here("src/directories.R"))

source(here("src/spatial.R"))

setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # set working directory to where this script is located
this_dir <- getwd()

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

impact_maps_dir <- file.path(biodiv_dir, "output")

```

```{r}
marine_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds")) %>%
     filter(sensitivity_scenario == "original") %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

terrestrial_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_terrestrial_impact.rds")) %>% 
   filter(sensitivity_scenario == "original") %>%
      dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total, total_hab_area)

all_overlap <- rbind(marine_overlap, terrestrial_overlap) %>%
    mutate(taxon = str_to_sentence(spp_type)) %>%
  mutate(taxon = case_when(
    taxon == "Terrestrial mammal" ~ "Terrestrial mammals",
    taxon == "Bird" ~ "Birds",
    taxon == "Fish" ~ "Finfish", 
    taxon == "Marine mammal" ~ "Marine mammals",
    taxon == "Marine plant" ~ "Marine plants",
    TRUE ~ taxon))

spp_total_hab <- all_overlap %>%
  distinct(sciname, spp_type, total_hab_area) %>%
  mutate(id = row_number())


ocean_impacts <- marine_overlap %>% 
  group_by(diet, fcr_type, allocation) %>%
  summarise(total_km2_impact = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(fcr_type == "regular", allocation == "economic") %>%
  mutate(type = "ocean")

terrestrial_impacts <- terrestrial_overlap %>%
  filter(!str_detect(ingredient, "fish meal|fish oil")) %>%
  group_by(diet, fcr_type, allocation) %>%
  summarise(total_km2_impact = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(fcr_type == "regular", allocation == "economic") %>%
  mutate(type = "terrestrial")


429026.4 + 123897.6 # total plant dominant
304212.8+210951.6 # total fish dominant
# so fish dominant has more total impacts what about averages per cell? 


all_impacts <- rbind(ocean_impacts, terrestrial_impacts)


```

## Results paragraph 1 


"  We find that the majority of species assessed are impacted to some degree by raw material production for the fish-dominant or plant-dominant feeds (n=42,471, 77.7% vs n=42,939, 78.6% respectively). "

```{r}
taxon_names <- all_overlap %>%
  distinct(spp_type, taxon)

spp_impacted <- all_overlap %>%
  filter(impact_total > 0) %>%
  group_by(taxon, diet, allocation, fcr_type) %>%
  summarise(spp_count_impacted = n_distinct(sciname)) %>%
  ungroup()

spp_impacted_economic <- spp_impacted %>% 
  filter(allocation == "economic", 
         fcr_type == "regular") %>%
  group_by(diet) %>%
  summarise(total_spp_impacted = sum(spp_count_impacted))

##  We find that the majority of species we assess are impacted to some degree by raw material production for the fish-dominant or plant-dominant feeds (n=42,473, 77.7% vs n=42,939, 78.6% respectively). 

aquamaps_spp <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  distinct(spp_type = taxon, sciname = species) %>%
  filter(spp_type != "polychaetes")

eyres_spp <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(spp_type, sciname = scientific_name) %>%
  mutate(sciname = tolower(sciname)) %>%
  mutate(spp_type = case_when(
    spp_type == "mammals" ~ "Terrestrial mammal",
    spp_type == "amphibians" ~ "Amphibians", 
    spp_type == "reptiles" ~ "Reptiles", 
    spp_type == "birds" ~ "Bird"
  ))

spp_assessed <- rbind(aquamaps_spp, eyres_spp) %>%
  left_join(taxon_names) %>%
  group_by(taxon) %>%
  summarise(spp_count_assessed = n_distinct(sciname)) %>%
  ungroup() %>%
  filter(!is.na(taxon))

spp_assessed_impacted <- spp_impacted %>%
  left_join(spp_assessed)

summary_info <- spp_assessed_impacted %>%
  group_by(diet, allocation, fcr_type) %>%
  summarise(total_assessed = sum(spp_count_assessed, na.rm = TRUE),
            total_impacted = sum(spp_count_impacted, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impacted = total_impacted/total_assessed)


marine_vs_terr <- spp_assessed_impacted %>%
  filter(diet == "plant-dominant", allocation == "economic", fcr_type == "regular") %>%
  mutate(type = ifelse(taxon %in% c("Amphibians", "Birds", "Terrestrial mammals", "Reptiles"), "terrestrial", "marine")) %>%
  group_by(type) %>%
    summarise(total_assessed = sum(spp_count_assessed, na.rm = TRUE),
            total_impacted = sum(spp_count_impacted, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impacted = total_impacted/total_assessed)

#  A tibble: 2 × 4
#   type        total_assessed total_impacted prop_impacted
#   <chr>                <int>          <int>         <dbl>
# 1 marine               22985          20538         0.894
# 2 terrestrial          31643          22401         0.708

spp_assessed %>% filter(taxon %in% c("Amphibians", "Reptiles", "Terrestrial mammals", "Birds")) %>% pull(spp_count_assessed) %>% sum() # 31643 assessed terrestial spp

```

"the mean proportion of species’ habitat impacted falls below 0.006% for 95% of cells "

```{r}
## read in plant-dominant, economic, regular
 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/economic_mean_prop.qs"))

all_df <-  qread(filepath)  %>%
  filter(value > 0) %>%
  filter(!is.na(value))

# test <- all_df %>% 
#   group_by(diet, fcr_type) %>%
#   summarise(avg = mean(value, na.rm = TRUE),
#             min = min(value, na.rm = TRUE),
#             max = max(value, na.rm = TRUE), 
#             median = median(value, na.rm = TRUE),
#             q_95 = quantile(value, c(0.05, 0.95))) %>%
#   ungroup()

options(scipen=999)

quantile(all_df %>% filter(fcr_type == "regular") %>% pull(value), c(0.05,0.95))
## both scenarios
#                    5%                   95% 
# 0.0000000000006245605 0.0000646685803076252 

quantiles_by_diet <- all_df %>%
  filter(fcr_type == "regular") %>%
  group_by(diet) %>%
  summarise(
    lower_quantile = quantile(value, 0.05),
    upper_quantile = quantile(value, 0.95)
  )

# Display the results
quantiles_by_diet

# A tibble: 2 × 3
#   diet           lower_quantile upper_quantile
#   <chr>                   <dbl>          <dbl>
# 1 fish-dominant        6.60e-13      0.0000497
# 2 plant-dominant       5.94e-13      0.0000848


max(all_df$value) # 0.1064724 # can dive into prep/03_prep_spp_habitats/08_process_mean_sd_summary.Rmd to find out more of whats happening in this cell. 

moll_template_land <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_land_xy.rds"))
moll_template_ocean <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_ocean_xy.rds"))

land_eez_rgns <- read.csv(file.path(biodiv_dir, "spatial/rgns_xy_all.csv")) %>% 
  dplyr::select(x, y, rgn_id)

maximum_location <- all_df %>%
  filter(value > 0.10) %>%
  left_join(land_eez_rgns, by = c("long" = "x", "lat" = "y")) %>%
 left_join(moll_template_land, by = c("long" = "x", "lat" = "y")) %>% # max value is ~0.22 on land in Norway
st_as_sf(coords = c("long", "lat"), crs = crs(moll_template)) # in Akershus county
mapview(maximum_location)

## lets look at differences between average cellular impacts on land and in ocean

land_impacts <- all_df %>%
  left_join(moll_template_land, by = c("long" = "x", "lat" = "y")) %>%
  filter(!is.na(cell_id)) %>% 
  group_by(diet, fcr_type) %>%
  summarise(avg_imp = mean(value, na.rm = TRUE)) %>%
  ungroup()

ocean_impacts <- all_df %>%
  left_join(moll_template_ocean, by = c("long" = "x", "lat" = "y")) %>%
  filter(!is.na(cell_id)) %>% 
  group_by(diet, fcr_type) %>%
  summarise(avg_imp = mean(value, na.rm = TRUE)) %>%
  ungroup()



```

## Results paragraph 2 

Comparing diets and impacts on land and ocean

"Intuitively, the fish-dominant diet exhibits larger impacts across all ocean pixels, most notably in high forage fish harvest regions such as the North Atlantic, Humboldt, and East China Sea (Figure 1b)."

"Similarly, the plant-dominant diet generally has larger land impacts, particularly in Argentina, Brazil, Chile, Russia, and Canada (Fig. 1b)."


```{r}
scenario_summary <- all_df %>%
  filter(value > 0) %>%
  group_by(diet, fcr_type) %>%
  summarise(cells_impacted = n()) %>%
  ungroup()


# The plant-dominant diet generally has larger land impacts, particularly in Argentina, Brazil, Chile, Poland and Canada (Figure 1b). 

moll_land <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_land_xy.rds"))

all_df_land <- all_df %>%
  left_join(moll_land, by = c("long" = "x", "lat" = "y")) %>%
  filter(!is.na(cell_id)) %>%
  filter(value > 0) %>%
    group_by(diet, fcr_type) %>%
  summarise(cells_impacted = n(),
                        impact_km2 = sum(value, na.rm = TRUE),
            mean_km2 = mean(value,na.rm = TRUE)) %>%
  ungroup() %>%
   mutate(land_cells_total = 2876998) %>%
  mutate(prop = cells_impacted/land_cells_total)


# Intuitively, the fish-dominant diet exhibits larger impacts across all ocean pixels, most notably in high forage fish harvest regions such as the North Atlantic, Humboldt region, and east China Sea (Figure 1b).
moll_ocean <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_ocean_xy.rds"))

all_df_ocean <- all_df %>%
  left_join(moll_ocean, by = c("long" = "x", "lat" = "y")) %>%
  filter(!is.na(cell_id)) %>%
    filter(value > 0) %>%
    group_by(diet, fcr_type) %>%
  summarise(cells_impacted = n(),
            impact_km2 = sum(value, na.rm = TRUE),
            mean_km2 = mean(value,na.rm=TRUE)) %>%
  ungroup() %>%
  mutate(ocean_cells_total = 3684240) %>%
  mutate(prop = cells_impacted/ocean_cells_total)


```

"The plant-dominant feed scenario impacts 37,760 km2 more total habitat globally than the fish-dominant (552,924 km2 vs. 515,164 km2, respectively). Additionally, the difference in the mean proportional area of habitat impacted per cell (a more meaningful metric that accounts for remaining habitat area 12,14) are greater on land than in the ocean in either scenario. The average cell impact is approximately 1.4 times higher in the ocean in the fish-dominant scenario, and on land, the plant-dominant diet results in over twice the average impact compared to the fish-dominant diet (~2.04x higher). Furthermore, the average land impacts from the fish-dominant diet are 12.8 times greater than the average fish-dominant ocean impacts (~18x larger than the plant-dominant ocean impacts), and the average plant-dominant land impacts are 36.7 times greater than their ocean counterparts. Thus, overall, greater agricultural dependence for feed provisioning appears to have further exacerbated impacts on terrestrial taxa."


```{r}


total_km2_diets <- all_overlap %>% 
  group_by(diet, allocation, fcr_type) %>%
  summarise(total_km2 = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic", 
         fcr_type == "regular")

# A tibble: 2 × 4
#   diet           allocation fcr_type total_km2
#   <chr>          <chr>      <chr>        <dbl>
# 1 fish-dominant  economic   regular    515164.
# 2 plant-dominant economic   regular    552924.

all_df_land %>% filter(fcr_type == "regular")

# A tibble: 2 × 7
#   diet           fcr_type cells_impacted impact_km2  mean_km2 land_cells_total  prop
#   <chr>          <chr>             <int>      <dbl>     <dbl>            <dbl> <dbl>
# 1 fish-dominant  regular          665433       36.3 0.0000545          2876998 0.231
# 2 plant-dominant regular          746371       82.8 0.000111           2876998 0.259

all_df_ocean %>% filter(fcr_type == "regular")
# # A tibble: 2 × 7
#   diet           fcr_type cells_impacted impact_km2   mean_km2 ocean_cells_total  prop
#   <chr>          <chr>             <int>      <dbl>      <dbl>             <dbl> <dbl>
# 1 fish-dominant  regular         1537181       6.55 0.00000426           3684240 0.417
# 2 plant-dominant regular         1537740       4.65 0.00000302           3684240 0.417

## ok so the mean impacts on land are bigger than on ocean, regardless of scenario comparisons! 

```

## Results paragraph 3

"Trade-offs among species between the scenarios become apparent when looking at impacts on individual species and aggregated by taxonomic groupings. The greatest average proportional global area of habitat impact at the taxa-mean are on terrestrial species under a plant-dominant scenario (Fig. 2). Furthermore, the greatest individual species global proportional impact is a reptile in the plant-dominant scenario (Liolaemus gravenhorstii, ~1.1% of its global habitat is impacted). Marine mammals have the greatest average proportional area of habitat impact under the fish-dominant scenario, however, the average impact for this taxa under the fish-dominant scenario is half that of terrestrial mammals under the plant-dominant scenario (0.00164% vs. 0.00312% respectively). Terrestrial species appear particularly sensitive to feed production activities -  terrestrial mammals, birds, and reptiles have a higher average proportional area of habitat impacted than even finfish under a fish-dominant scenario. This is despite over 50% of the fish-dominant feed being derived from marine materials. With this in mind, and wild-sourced fishmeal and oil already minimised by necessity, a priority for future feed development is to minimise the biodiversity impacts of feed sourcing on land."

```{r}
taxon_names <- all_overlap %>%
  distinct(spp_type, taxon)

spp_impacted <- all_overlap %>%
  filter(impact_total > 0) %>%
  group_by(taxon, diet, allocation, fcr_type) %>%
  summarise(spp_count_impacted = n_distinct(sciname)) %>%
  ungroup()

aquamaps_spp <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  distinct(spp_type = taxon, sciname = species)
  
eyres_spp <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(spp_type, sciname = scientific_name) %>%
  mutate(sciname = tolower(sciname)) %>%
  mutate(spp_type = case_when(
    spp_type == "mammals" ~ "Terrestrial mammal",
    spp_type == "amphibians" ~ "Amphibians", 
    spp_type == "reptiles" ~ "Reptiles", 
    spp_type == "birds" ~ "Bird"
  ))

spp_assessed <- rbind(aquamaps_spp, eyres_spp) %>%
  left_join(taxon_names) %>%
  group_by(taxon) %>%
  summarise(spp_count_assessed = n_distinct(sciname)) %>%
  ungroup() %>%
  filter(!is.na(taxon))

sum(spp_assessed$spp_count_assessed) # 54628



spp_assessed_plot_df <- spp_assessed %>%
  mutate(spp_count_str = glue("n = {spp_count_assessed}"))

all_overlap_df <- all_overlap %>%
  rename(fcr = fcr_type) %>%
        separate_wider_delim(ingredient, delim = "_", names = c("raw_material", "ingredient")) %>% 
  mutate(raw_material = str_to_sentence(raw_material), 
         ingredient = str_to_sentence(ingredient),
         allocation = str_to_sentence(allocation),
         diet = str_to_sentence(diet),
         fcr = str_to_sentence(fcr)) %>%
    filter(allocation == "Economic",
           fcr == "Regular")

length(unique(all_overlap_df$sciname)) # 54625

summary_df_spp <- all_overlap_df %>%
  group_by(sciname, diet, fcr, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(impact_total > 0) %>%
  left_join(spp_assessed_plot_df) %>%
  left_join(spp_total_hab %>% dplyr::select(-spp_type, -id)) %>%
  mutate(prop_impact = impact_total/total_hab_area)  %>%
  group_by(taxon, diet, fcr) %>%
  mutate(mean_prop = mean(prop_impact,na.rm=TRUE),
         median_prop = median(prop_impact,na.rm=TRUE)) %>%
  ungroup()


mean_df <- summary_df_spp %>%
  distinct(diet, taxon, mean_prop) %>%
  pivot_wider(values_from = mean_prop, names_from = diet) %>%
  mutate(`Plant-dominant`/`Fish-dominant`)

mean_df %>% filter(taxon == "Terrestrial mammals") %>% pull(`Fish-dominant`) # 0.00001449416
mean_df %>% filter(taxon == "Finfish") %>% pull(`Fish-dominant`) # 0.0000134733

0.00001449416 > 0.0000134733 # TRUE


taxon_hab_area <- spp_total_hab %>%
  dplyr::select(-sciname, -id) %>%
  group_by(spp_type) %>%
  summarise(total_hab_area = sum(total_hab_area, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_names) %>%
  dplyr::select(-spp_type)
  

summary_df_all <- all_overlap_df %>%
  group_by(diet, fcr, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_hab_area) %>%
  mutate(prop_impact = impact_total/total_hab_area)

quant_95 <- quantile(summary_df_spp$prop_impact, 0.95)


summary_df_spp_out <- summary_df_spp %>%
  filter(prop_impact >= quant_95) %>%
  group_by(diet, fcr, allocation, taxon) %>%
  mutate(spp_outliers = n_distinct(sciname)) %>%
  ungroup()  %>%
  mutate(prop = spp_outliers/spp_count_assessed)

# Moreover, of the terrestrial mammal species assessed (n=5502), 8.1% are above the 95th percentile proportion of total habitat impacted with the plant-dominant scenario, the highest percentage of species falling in this category among taxa and scenarios.
summary_df_spp_out %>% filter(diet == "Plant-dominant", taxon == "Terrestrial mammals") %>% pull(spp_outliers) %>% unique()

# 8.3% terrestrial mammals >95th quantile 
```


## Results paragraph 4 

"Greater average impacts to terrestrial species can be explained by a number of factors. One is that the average habitat size of impacted terrestrial species is 12.5 times smaller than marine, yielding greater proportional impact for every square kilometre appropriated. Further, from available data, terrestrial species appear more sensitive to crop production than marine species to fishing activities. For 88% of marine species assessed, heavily fished areas can still represent marginal habitat, whereas agricultural production renders habitat completely unsuitable for 81% of the terrestrial species we assess (fishing activities render habitat completely unsuitable for only 2% of marine species)."


```{r}

# One explanation is that the average habitat size of impacted terrestrial species is 12.5 times smaller than marine, yielding greater proportional impact for every square kilometre appropriated.

marine_hab_areas <-  marine_overlap %>%
  filter(impact_total > 0) %>%
  distinct(sciname, total_hab_area)

mean(marine_hab_areas$total_hab_area) # 6577977

terr_hab_areas <- terrestrial_overlap %>%
  filter(impact_total > 0) %>%
  distinct(sciname, total_hab_area)

mean(terr_hab_areas$total_hab_area) # 523836.1


6577977/523836.1 # 12.55732 marine habitat areas are ~12.5x larger than terrestrial on average


# Further, from available data, terrestrial species appear more sensitive to crop production than marine species to fishing activities. For most marine species in our analysis, heavily fished areas can still represent marginal habitat (88% of marine species), whereas agricultural production renders habitat completely unsuitable for 81% of the terrestrial species we assess (compared with fishing activities rendering habitat completely unsuitable for only 2% of marine species)."


vuln_values <- all_overlap %>%
  distinct(sciname, ingredient, vulnerability) %>%
  mutate(type = ifelse(str_detect(ingredient, "fish"), "marine", "terrestrial")) %>%
  group_by(type) %>%
  summarise(mean_vuln = mean(vulnerability)) %>%
  ungroup()

ag_suitability <- readRDS(here("prep/03_prep_spp_habitats/int/terrestrial_spp_habitat_suitability.rds"))

test <- ag_suitability %>% 
  mutate(sensitivity = 1 - cropland_suitability) %>%
  group_by(sensitivity) %>%
  summarise(count_sens = n())

25487/sum(test$count_sens) # 0.8066527 ~81% of terrestrial species are fully sensitive

vuln_taxa <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  unite("ingredient", c(fish_type, ingredient), sep = "_") %>%
  mutate(ingredient = case_when(
    ingredient == "trimmings fish_fish meal" ~ "trimmings fish_fish meal, cut offs",
    ingredient == "trimmings fish_fish oil" ~ "trimmings fish_fish oil, cut offs",
    TRUE ~ ingredient
  )) %>%
  distinct(species,ingredient, vuln_quartile) %>%
  filter(species %in% unique(marine_overlap$sciname))

test2 <- vuln_taxa %>%
  group_by(species) %>% 
  summarise(sensitivity = mean(vuln_quartile)) %>%
  ungroup() %>%
  group_by(sensitivity) %>% 
  summarise(count_sens = n())

412/sum(test2$count_sens) # 0.01789049 ==> ~2% of marine species are full sensitive.. 

(20265 + 4 + 1)/sum(test2$count_sens) # 0.8801945 vast majority (~88% are marginal)


terr_vuln_spp <- ag_suitability %>% 
  filter(cropland_suitability != 1)
length(terr_vuln_spp$species) # 26413 terrestrial species are sensitive to cropland


marine_vuln_spp <- vuln_taxa %>%
  group_by(species) %>%
  summarise(total_vuln = sum(vuln_quartile, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(total_vuln != 0)

length(marine_vuln_spp$species) # 20682 marine species are sensitive to harvest

(20682 + 26413)/(23029 + 31596) # 0.8526659 are vulnerable in some capacity

# Average impact differences across terrestrial and marine taxa mask information about the number of species impacted. A similar number of species are impacted on land and sea across both feed scenarios. Of the 31,643 terrestrial species we assessed, 22,401 (71%) were impacted by feed crop production to some extent. In the marine environment, 20,538 of the 22,985 assessed species (89%) were impacted by fishing for fishmeal and oil.

```

## Results paragraph 5 

Our assessment enables impact comparisons across raw materials and taxa for different feed formulations, and reveals that different dependence on marine and terrestrial materials does not predictably shift the burden of impacts. For example, forage fish harvest accounts for the largest absolute impact area in the fish-dominant scenario (300,536 km2), and soybean production accounts for the largest in the plant-dominant scenario (241,910 km2). However, when observing the proportion of total global taxa habitat area impacted, we find that soybean predominates in both scenarios (Fig. 3, Fig. 1). This is due to a high inclusion of soybean ingredients (% of formulation) in both feed scenarios, and smaller habitat areas of terrestrial taxa, as discussed above. As expected, nearly all of the impact on marine species comes from forage fish-derived FMFO. Our model allocates little marine disturbance to the FMFO from trimmings as our allocation approach assumes that it is the edible portions of fish valued for human consumption, rather than the subsequent processing waste, that is the main driver of fishing activities and their environmental impact (see Methods for the allocation approach used for biomass estimation). Within marine taxa, macrophytes and benthic invertebrates (i.e., echinoderms, sponges, molluscs, arthropods) appear most impacted from forage fish harvest. These benthic or benthic-adjacent species are often caught as bycatch in gears like bottom trawling due to their water column positions (~10% of catch in our assessment comes from destructive gears). By analysing the average impacts across taxa in each cell, we find that soybean and rapeseed have the most, and nearly equal average impacts globally (Supplementary Fig. 4). That is, while soybeans may contribute the most to the total and proportional global impact (~3.3x and ~2.3x more than rapeseed, respectively), rapeseed causes nearly equal habitat displacement locally in the average cell.

```{r}
all_overlap_df %>% 
  group_by(raw_material, diet) %>%
  summarise(impact_km2 = sum(impact_total, na.rm = TRUE)) %>%
  arrange(-impact_km2)

# A tibble: 18 × 3
# Groups:   raw_material [9]
#    raw_material   diet           impact_km2
#    <chr>          <chr>               <dbl>
#  1 Forage fish    Fish-dominant     300536.
#  2 Soybean        Plant-dominant    241910.
#  3 Soybean        Fish-dominant     149250.
#  4 Forage fish    Plant-dominant    122400.
#  5 Rapeseed       Plant-dominant     73246.
#  6 Wheat          Fish-dominant      52112.
#  7 Pulses         Plant-dominant     45239.
#  8 Wheat          Plant-dominant     40869.
#  9 Sunflower      Plant-dominant     18421.
# 10 Cropsnes       Plant-dominant      8934.
# 11 Pulses         Fish-dominant       7267.
# 12 Trimmings fish Fish-dominant       3676.
# 13 Maize          Fish-dominant       2322.
# 14 Trimmings fish Plant-dominant      1497.
# 15 Maize          Plant-dominant       406.
# 16 Cropsnes       Fish-dominant          0 
# 17 Rapeseed       Fish-dominant          0 
# 18 Sunflower      Fish-dominant          0 


spp_impacted_by_material <- all_overlap_df %>%
  filter(impact_total > 0) %>%
  group_by(raw_material, diet) %>%
  summarise(spp_impacted = list(unique(sciname)))

spp_total_hab_by_material <- spp_impacted_by_material %>%
  rowwise() %>%
  mutate(total_hab_area = sum(spp_total_hab %>%
                                filter(sciname %in% spp_impacted) %>%
                                pull(total_hab_area), na.rm = TRUE)) %>%
  select(raw_material, diet, total_hab_area)

# Step 1: Summarise total habitat area per raw material
spp_total_hab_by_material <- spp_impacted_by_material %>%
  rowwise() %>%
  mutate(total_hab_area = sum(spp_total_hab %>%
                                filter(sciname %in% spp_impacted) %>%
                                pull(total_hab_area), na.rm = TRUE)) %>%
  select(raw_material, diet, total_hab_area)

# Step 2: Summarise total impact_km2 per raw material
total_impact_by_material <- all_overlap_df %>%
  group_by(raw_material, diet) %>%
  summarise(impact_km2 = sum(impact_total, na.rm = TRUE))

# Step 3: Merge the two datasets
impact_hab_combined <- total_impact_by_material %>%
  left_join(spp_total_hab_by_material, by = c("raw_material", "diet"))

# Step 4: Divide total impact by total habitat area per raw material
impact_hab_combined <- impact_hab_combined %>%
  mutate(impact_per_hab_area = impact_km2 / total_hab_area) %>%
  arrange(-impact_per_hab_area)

# A tibble: 18 × 5
# Groups:   raw_material [9]
#    raw_material   diet           impact_km2 total_hab_area impact_per_hab_area
#    <chr>          <chr>               <dbl>          <dbl>               <dbl>
#  1 Soybean        Plant-dominant    241910.    10769675500        0.0000225   
#  2 Soybean        Fish-dominant     149250.    10769675500        0.0000139   
#  3 Rapeseed       Plant-dominant     73246.     7378281300        0.00000993  
#  4 Wheat          Fish-dominant      52112.    10792393500        0.00000483  
#  5 Pulses         Plant-dominant     45239.    11677240900        0.00000387  
#  6 Wheat          Plant-dominant     40869.    10792393500        0.00000379  
#  7 Forage fish    Fish-dominant     300536.   135401073300        0.00000222  
#  8 Sunflower      Plant-dominant     18421.     9677818200        0.00000190  
#  9 Forage fish    Plant-dominant    122400.   135401073300        0.000000904 
# 10 Cropsnes       Plant-dominant      8934.    11709295400        0.000000763 
# 11 Pulses         Fish-dominant       7267.    11677240900        0.000000622 
# 12 Maize          Fish-dominant       2322.    11677143300        0.000000199 
# 13 Maize          Plant-dominant       406.    11677143300        0.0000000348
# 14 Trimmings fish Fish-dominant       3676.   126560378400        0.0000000290
# 15 Trimmings fish Plant-dominant      1497.   126560378400        0.0000000118
# 16 Cropsnes       Fish-dominant          0              NA       NA           
# 17 Rapeseed       Fish-dominant          0              NA       NA           
# 18 Sunflower      Fish-dominant          0              NA       NA   

## proportion impacted for spp which soybean impacts is greater than all else

# Within marine species, we find that macrophytes and less mobile benthic invertebrates (i.e., echinoderms, cnidaria, sponges, molluscs, arthropods) with smaller habitat areas are the most impacted from forage fish harvest.

# These benthic or benthic-adjacent species are often caught as bycatch from gears like trawling and traps due to their water column positions (~10% of catch in our assessment comes from these destructive gears).

test <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>% 
  filter(taxon %in% c("arthropods", "molluscs", "echinoderms", "cnidaria", "sponges", "corals", "Marine plant"))

test2 <- test %>% group_by(depth_position) %>% summarise(n())

test3 <- test %>% group_by(depth_position, taxon, catch_type) %>% summarise(n())

# lets look at gears used for forage fish catch

catch <- readRDS(file.path(watson_dir, "v5.0/int/all_catch_foragefish.rds"))

test_catch <- catch %>% 
  filter(FOFM_catch > 0) %>%
  group_by(water_col_position, VBDesc) %>%
  summarise(forage_catch = sum(FOFM_catch,na.rm = TRUE)) %>%
  ungroup()

sum(catch$FOFM_catch) # 24765797

unique(test_catch$VBDesc)

test_catch %>%
  filter(VBDesc %in% c("bottom trawls", "conical and drum-like traps", "trapping gear", "traps", "by diving", "explosives")) %>%
  pull(forage_catch) %>%
  sum()

# 3339241/24765797

trimmingsfish <- readRDS(file.path(watson_dir, "v5.0/int/all_catch_trimmingsfish.rds")) %>%
  filter(trimmings_catch > 0) %>%
  group_by(water_col_position, VBDesc) %>%
  summarise(trimmings_catch = sum(trimmings_catch,na.rm = TRUE)) %>%
  ungroup()

sum(trimmingsfish$trimmings_catch) # 18282816

unique(trimmingsfish$VBDesc)

trimmingsfish %>%
  filter(VBDesc %in% c("bottom trawls", "conical and drum-like traps", "trapping gear", "traps", "by diving", "explosives")) %>%
  pull(trimmings_catch) %>%
  sum()

(1129209 + 3339241)/(18282816 +24765797) # 10% of catch comes from destructive benthic gears


terr_taxa_avg <- terrestrial_overlap %>% 
  filter(impact_total > 0) %>%
  distinct(sciname, spp_type, total_hab_area) %>%
  group_by(spp_type) %>%
  summarise(avg_hab = mean(total_hab_area, na.rm = TRUE),
            total_hab = sum(total_hab_area, na.rm = TRUE)) %>%
  ungroup()

# # A tibble: 4 × 3
#   spp_type           avg_hab  total_hab
#   <chr>                <dbl>      <dbl>
# 1 Amphibians         166603.  798028100
# 2 Bird               804522. 5403171500
# 3 Reptiles           353182. 2256831500
# 4 Terrestrial mammal 729385. 3252325600

# Similarly, amphibians have the smallest global average and total habitat areas for terrestrial taxa and are the most heavily impacted, largely from soybean, wheat, and pulse production (Figure 3).

```

## Results paragraph 6

"These results highlight raw materials that have disproportionate environmental impacts compared to their use in feed. For example, soybeans represent 51% and 41% of the species area of total habitat impacted, but only 17% and 22% of the feed formulation under the fish-dominant and plant-dominant scenarios, respectively. For all the other materials, impacts are nearly equal or lower percentages than their contribution to feed formulation (Supplementary Fig. 6). This discrepancy is likely due to soybean harvest overlapping with highly biodiverse regions (e.g., Argentina, Brazil), and species which are particularly sensitive to cropland 23,31. By identifying high-impact raw materials, strategies can be developed for sustainable sourcing and long-term goals for novel ingredient inclusion (e.g., insect meal), including minimising the overall ecological footprint of feed production and preserving vulnerable species."


Look at "efficiency" of raw materials; how does proportional contribution of impact differ from proportional contribution of feed formulation?

1. Look at proportional contribution of total overall impacts per raw material


```{r}

codes <- read.csv(here("prep/02_feed/data/ingredient_spam_fao_codes.csv")) %>%
  dplyr::select(raw_name = source_ingredient, raw_material = raw.material)

diets_raw_mat <- read.csv(here("prep/02_feed/data/aquaculture_diet_composition.csv")) %>%
  distinct(raw_name, prop_diet, diet) %>%
  left_join(codes) %>%
  mutate(raw_material = case_when(
    str_detect(raw_name, "cut offs") ~ "Trimmings fish", 
    str_detect(raw_name, "pea flour|guar|pea pro|faba") ~ "Pulses",
    str_detect(raw_name, "forage fish") ~ "Forage fish",
    str_detect(raw_name, "coconut|linseed") ~ "Cropsnes",
    TRUE ~ str_to_title(raw_material)
    
  )) %>%
  group_by(raw_material, diet) %>%
  summarise(total_diet_contribution = sum(prop_diet, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(!is.na(raw_material)) %>%
  mutate(diet = str_to_sentence(diet))

summary_df_barcharts <- all_overlap_df %>%
  group_by(diet, fcr, raw_material, allocation, taxon) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(taxon_hab_area) %>%
  mutate(prop_impact = impact_total/total_hab_area)

summary_df_stats <- summary_df_barcharts %>%
   group_by(diet, fcr, raw_material, allocation) %>%
  summarise(prop_total_impact = sum(prop_impact),
            total_km2 = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr, allocation) %>% 
  mutate(total_impact = sum(prop_total_impact)) %>%
  mutate(prop_attributed = prop_total_impact/total_impact) %>%
  left_join(diets_raw_mat) %>%
  mutate(delta = prop_attributed - total_diet_contribution) %>%
  mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material))
  
## wow ok some interesting results! Soybean is highly inefficient! Rapeseed doesn't appear to be that bad...
# In the fish-dominant diet: 
#  - soybean accounts for 17% of the feed, however, it accounts for ~51% of the impacts. 
#  - forage fish ingredients account for 42% of the fish dominant feed and only ~41% of the impacts. 
#  - wheat is similarly efficient in the fish dominant diet; 
# 
# For the plant-dominant diet:
#  - Soybeans are 22% of the feeds but account for ~41% of the impacts.
#  - Forage fish are 17% of the feed and 17% of the impacts. 
#  - Wheat is 16% of the feed and 11% of the impacts
#  - Pulses are 12% of the feed and 11% of the impacts
#  - Rapeseed is 18% of the feed and 10% of the impact

## lets look at average impacts contributions

# read in average raw material impacts - country level... unfortunately will have to take a mean here 
csv_files <- list.files(file.path(impact_maps_dir, "csvs/by_material_global_summary"), pattern = ".qs", full.names = TRUE)

list_of_dfs <- lapply(csv_files, qread)


combined_impact_df <- do.call(rbind, list_of_dfs) %>%
  mutate(raw_material = str_to_sentence(raw_material), 
         allocation = str_to_sentence(allocation),
         diet = str_to_sentence(diet),
         fcr = str_to_sentence(fcr))  %>%
 mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material))

material_stats_1 <- combined_impact_df %>%
    filter(allocation == "Economic") %>%
  filter(fcr == "Regular") 

  
summary_df_stats_diet <- material_stats_1 %>%
    group_by(diet) %>% 
  mutate(total_impact = sum(mean_prop_weighted)) %>%
  mutate(prop_attributed = mean_prop_weighted/total_impact) %>%
  left_join(diets_raw_mat) %>%
  mutate(delta = prop_attributed - total_diet_contribution) %>%
  mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material)) %>%
  mutate(delta_abs = abs(delta))



ggplot(summary_df_stats_diet, aes(x = prop_attributed, y = total_diet_contribution, color = raw_material)) +
  geom_point(size = 3) +
  facet_wrap(~ diet) +
  labs(color = "",
       x = "Average Impact Contribution",
       y = "Diet Contribution") +
  theme_bw() +
  theme(legend.position = "bottom") +
  geom_abline() +
  scale_color_manual(values = pal_mats)


```

 
## Results paragraph 7
 - See case study script


## Random stuff not in MS 

How much raw material does Norway produce and use in country for salmon aquafeeds? 

```{r}
ingredient_production_location <- read.csv(here("prep/02_feed/data/demand/ingredient_demand_production_locations.csv"))

norway_incountry <- ingredient_production_location %>%
  filter(
         iso3c_consuming == "NOR") %>%
  group_by(iso3c_consuming, iso3c_producing, diet, fcr_type, GAEZ_category) %>%
  summarise(total_tonnes_traded = sum(producing_ingredient_consumed_tonnes, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr_type) %>%
  mutate(total_tonnes_consumed = sum(total_tonnes_traded, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_consumed = total_tonnes_traded/total_tonnes_consumed)

```

How much salmon aquaculture production does NOR produce? 

```{r}

production <- readRDS(here("data/tidy_data/production-data/aquaculture_production_tidy.rds")) 


salmon_production <- production %>%
  filter(species == "Atlantic salmon" & year == 2020,
         value > 0) %>%
  mutate(species = "salmon")

1388433.84/sum(salmon_production$value) # 51%

```


How many species have both marine and terrestrial impacts? 

```{r}

test <- all_overlap %>%
  filter(spp_type %in% c("Bird", "Amphibians", "Reptiles")) %>%
  filter(impact_total > 0) %>%
  filter(allocation == "economic", fcr_type == "regular") %>%
  mutate(ingredient_type = ifelse(str_detect(ingredient, "forage fish|trimmings fish"), "marine", "terrestrial")) %>%
  pivot_wider(names_from = ingredient_type, values_from = impact_total)

unique(test$sciname) # 46 amphibians, reptiles, or birds which have marine impacts

## do any of them have both??
test2 <- test %>%
  group_by(sciname, spp_type, diet) %>%
  summarise(marine = sum(marine, na.rm = TRUE),
            terrestrial = sum(terrestrial, na.rm = TRUE)) %>%
  ungroup() %>% 
  filter(marine > 0, terrestrial > 0) ## nope, none of them have both

```


How much bigger are marine habitats than terrestrial habitats?


```{r}

marine_hab_areas <-  marine_overlap %>%
  filter(impact_total > 0) %>%
  distinct(sciname, total_hab_area)

mean(marine_hab_areas$total_hab_area) # 6577977

terr_hab_areas <- terrestrial_overlap %>%
  filter(impact_total > 0) %>%
  distinct(sciname, total_hab_area)

mean(terr_hab_areas$total_hab_area) # 523836.1


6577977/523836.1 # 12.55732 marine habitat areas are ~12.5x larger than terrestrial on average


## what about impacts 
test <- marine_overlap %>%
  filter(
          impact_total > 0,
         allocation == "economic") %>%
  group_by(sciname, diet, fcr_type, allocation) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(marine_hab_areas) %>%
  mutate(prop_impact = impact_total/total_hab_area) %>% 
  group_by(diet, fcr_type) %>%
  summarise(mean_prop = mean(prop_impact, na.rm = TRUE)) %>% 
  ungroup()

## mean plant-dominant marine spp prop impact is 
## filter 0s: 0.000003810617

test2 <- terrestrial_overlap %>%
  filter(
          impact_total > 0,
         allocation == "economic") %>%
  group_by(sciname, diet, fcr_type, allocation) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(terr_hab_areas) %>%
  mutate(prop_impact = impact_total/total_hab_area) %>% 
  group_by(diet, fcr_type) %>%
  summarise(mean_prop = mean(prop_impact, na.rm = TRUE)) %>% 
  ungroup()

# mean plant-dominant terrestrial spp prop impact is 
# 0.00002844402

## terrestrial plant-dominant impacts are on average 7.5x more than marine plant-dominant impacts
0.00002844402/0.000003810617

## terrestrial fish-domiannt impacts are on average 1.4x more than marine fish-dominant impacts
0.000009356426 # marine fd
0.00001272571 # terr fd

0.00001272571/0.000009356426 # 1.360104



```
