==> analysis/07_Fig4_case_study.Rmd <==
---
title: "Case study for publication"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary

This script has code which produces our example case study figure, Figure 4: 

> Figure 4. Impact hotspot map of central Chile, under the plant-dominant feed scenario. (A) Cells with average impacts greater than or equal to the 95th quantile. (B) The raw material which contributes most to the impact in each hotspot cell. (C) The proportion of impacted species in each cell which is at-risk of extinction. (D) Reference map, using Open Street Map (https://www.openstreetmap.org/). 

For reference, I had to combine each of the panels in an image editor (I used Inkscape to combine), because I could not get it to look good in R. 

## Setup

```{r setup, include=FALSE}

## load libraries
library(tidyverse)
library(here)
library(data.table)
library(terra)
library(strex)
library(patchwork)
library(cowplot)
library(glue)
library(RColorBrewer)
library(ggpubr)
library(qs)
library(parallel)
library(doParallel)
library(sf)
library(rnaturalearth)
library(ggspatial)

# call necessary functions
source(here("src/directories.R"))
source(here("src/spatial.R"))

# establish filepaths
biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")
impact_maps_dir <- file.path(biodiv_dir, "output")
this_dir <- here("analysis/")

# read in shapefile with country borders
countries_shp <- ne_countries(scale = 10, returnclass = "sf") |> 
  st_transform(crs(moll_template)) # read in countries shapefile

borders_shp <- ne_countries(scale = 110, returnclass = "sf") |> 
  st_transform(crs(moll_template)) # read in countries shapefile

globe_bbox <- rbind(c(-180, -90), c(-180, 90), 
                      c(180, 90), c(180, -90), c(-180, -90)) 

globe_border <- st_polygon(list(globe_bbox)) %>%
  st_sfc(crs = 4326) %>%
  st_sf(data.frame(rgn = 'globe', geom = .)) %>%
  smoothr::densify(max_distance = 0.5) %>%
  st_transform(crs = crs(moll_template))


# establish color palettes
pal_mats <- c("Soybean" = "#328C41",
              "Trimmings fish" = "#8FBAED",
              "Sunflower" = "#4277A9",
              "Wheat" = "#FFE96D",
              "Rapeseed" = "#AC9900",
              "Pulses" = "#79E0DA",
              "Forage fish" = "#3DB1AB",
              "Maize" = "#0A6B66",
              "Other crops" = "#BC94C6")

quantile_pal <- c(
    "0-24" = "#D4F7F0", 
    "25-49" = "#538AC6", 
    "50-74" = "#1E438F", 
    "75-94" = "#F8D422", 
    "95-99" = "#E95D08", 
    "≥99" = "#781211"
)

iucn_pal <- c("NE" = "#808080",
              "DD" = "#808080",
              "LC" = "#46A440",
              "NT" = "#BDC72D",
              "VU" = "#F8D422",
              "EN" = "#E95D08",
              "CR" = "#781211")

```

Explore the cell with the highest impacts:
 - in Norway, plant-dominant 11% average habitat area impacted
 - in Norway, fish-dominant 9% average habitat area impacted
 
```{r}
# read in file with country ids associated to cells, so that we can know which country impacts are in
land_eez_rgns <- read.csv(file.path(biodiv_dir, "spatial/rgns_xy_all.csv")) %>% 
  dplyr::select(x, y, rgn_id)

region_ids <- read.csv(here("data/spatial/output/region_ids.csv"))

## read in economic allocation impacts for all scenarios
 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/economic_mean_prop.qs"))

all_df <-  qread(filepath)  %>%
  filter(value > 0) %>%
  filter(!is.na(value))

max_scen <- all_df %>%
  group_by(diet, fcr_type) %>%
  summarise(max_val = max(value, na.rm = TRUE)) %>%
  ungroup()

## here are the max impacts in a cell for each scenario
# # A tibble: 2 × 3
#   diet           fcr_type max_val
#   <chr>          <chr>      <dbl>
# 1 fish-dominant  regular   0.106 
# 2 plant-dominant regular   0.0925

## filter dataset for cells with >0.09 to determine where these cells are
max_cell <- all_df %>%
  filter(value > 0.09) # 786772 cell_id

## lets find the exact country of these impacts, and if they are on land or in the ocean
moll_template_land <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_land_xy.rds")) %>%
  mutate(type = "land")
moll_template_ocean <- readRDS(here("prep/03_prep_spp_habitats/data/spatial/moll_template_ocean_xy.rds")) %>%
  mutate(type = "ocean")

moll_template_all <- rbind(moll_template_land, moll_template_ocean)

maximum_location <- all_df %>%
  # filter(value >  0.09) %>% # filter for large impacts
  left_join(land_eez_rgns, by = c("long" = "x", "lat" = "y")) %>% # join with country ids
  left_join(moll_template_all, by = c("long" = "x", "lat" = "y")) # join with df denoting whether cells are land or ocean

# All max values in each scenario are in the same cell, on land, in Norway: cell_id = 786772
# side note - the maximum ocean cells are also in Norway
# another side note all the cells with impacts >0.09 are in Norway. 
# another side note; Norway has highest impacts all the way til 0.03307191, then rgn_id = 224 shows up with an land cell having an impact of 0.03307191; this is Finland

## how many spp impacted in that cell? 
 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/economic_nspp.qs"))

nspp_df <-  qread(filepath) %>%
  left_join(moll_template_xy, by = c("long" = "x", "lat" = "y")) %>%
  dplyr::select(cell_id, diet, fcr_type, nspp_impacted = value) %>%
  filter(cell_id == 786772) # 39 spp impacted in all scenarios

## how many species are actually in that cell? 

all_richness <- rast(list.files(here("data/spatial/output"), pattern = "richness", full.names = TRUE)) %>% 
    app(., "sum", na.rm = TRUE) %>%
  as.data.frame(., xy = TRUE) %>%
  left_join(moll_template_xy) %>%
  dplyr::select(cell_id, sum_spp = sum)

max_impacts <- maximum_location %>% 
  left_join(all_richness) %>%
  left_join(nspp_df) %>%
 filter(cell_id == 786772)

## ok so, 39 spp impacted, 97 spp present in the cell with maximum impacts

```

Ok let's dive a bit deeper by looking into larger areas of impact, rather than just one cell.
What species are impacted in largest impact cells:
 - grab any cells that fall into the 95-99 or >= 99 quantile categories; cell list saved in 02_publication_maps.Rmd
 - I've done this separately for marine and terrestrial, and chunked into each scenario, just incase they get too large and unwieldy.
 - additionally, I've only filtered for these countries, to cut down on the size of the data: "NOR", "CAN", "ARG", "CHL", "AUS"
 
 - ONLY NEEDS TO BE RUN ONE TIME 
 
```{r}

large_impact_cells <- readRDS(here("analysis/int/large_impact_cells.rds")) 

## lets only do this for regular and economic scenarios
diet_types <- c("fish-dominant", "plant-dominant")
fcr_types <- c("regular")
allocation_types <- c("economic")


max_impact_df <- data.frame(species_full = NA, cell_id = NA, cropland_suitability = NA, spp_type = NA, filepath = NA,
                            species_id = NA, harvest = NA, hab_area = NA, impact_km2 = NA, source_file = NA,
                            raw_material = NA, diet = NA, allocation = NA, fcr = NA)

for(diet in diet_types){
  for(fcr in fcr_types){
    for(allocation in allocation_types){

# allocation = "economic"
# diet = "plant-dominant"
# fcr = "regular"

all_files <- list.files(file.path(biodiv_dir, "output/impact_maps_by_spp_ingredient_lists"), pattern = glue(".*_{diet}_{fcr}_original_.*_{allocation}"), full.names = TRUE)

to_remove <- grep("fish oil|fish meal", all_files)
all_files <- all_files[-to_remove]


all_files_list <- lapply(all_files, qread)

## now filter for specific cell_id and see what is impacted...

max_cell_impacts <- all_files_list %>% 
  map(~ .x %>%
        map(~ filter(.x, cell_id %in% c(large_impact_cells))))  # filter for any cells that are in the cells that have an average impact in the 95-99 or >=99th quantiles, in the countries listed: "NOR", "CAN", "ARG", "CHL", "AUS"


for (i in seq_along(max_cell_impacts)) {
  for (j in seq_along(max_cell_impacts[[i]])) {
    max_cell_impacts[[i]][[j]]$source_file <- all_files[i]
  }
}

max_cell_impacts_fin <- max_cell_impacts %>%
  bind_rows() %>%
  mutate(raw_material = str_before_first(str_after_nth(str_after_last(source_file, "/"), "_", 3), "_")) %>%
  mutate(diet = diet, 
         allocation = allocation, 
         fcr = fcr)

qsave(max_cell_impacts_fin, glue(here("analysis/int/large_cell_impacts_terrestrial_{diet}_{fcr}_{allocation}.qs"))) # save as qs file to save space


    }
  }
}



```

Need to do the same thing for marine impacts. I am chunking this just in case the data gets too large and unwieldy. 

```{r}
diet_types <- c("plant-dominant")
fcr_types <- c("regular")
allocation_types <- c("economic")


for(diet in diet_types){
  for(fcr in fcr_types){
    for(allocation in allocation_types){
      
      # diet = "fish-dominant"
      # fcr = "regular"
      # allocation = "economic"
      
max_impact_df <- data.frame(species_full = NA, cell_id = NA, cropland_suitability = NA, spp_type = NA, filepath = NA,
                            species_id = NA, harvest = NA, hab_area = NA, impact_km2 = NA, source_file = NA,
                            raw_material = NA, diet = NA, allocation = NA, fcr = NA)      
      
for(chunk_id in 1:13){
        
       # chunk_id = 4
        
        files <- list.files(file.path(biodiv_dir, "int/marine_aggregation/sub_chunk_dfs/"), pattern = glue(".*_{diet}_{fcr}_original_.*_{allocation}_chunk_{chunk_id}.qs"), full.names = TRUE)
        
        impacts_chunk <- lapply(files, qread) %>%
          bind_rows() %>%
          filter(!is.na(cell_id)) %>%
          filter(cell_id %in% large_impact_cells) %>%
  as.data.frame() %>% 
  mutate(diet = diet, 
         allocation = allocation, 
         fcr = fcr)

        qsave(impacts_chunk, glue(here("analysis/int/chunked/large_cell_impacts_marine_{diet}_{fcr}_{allocation}_chunk_{chunk_id}.qs")))


      }
                 
    }
  }
} 

## seems that the chunking may have been unnecessary


fd_impacts_list <- list.files(here("analysis/int/chunked/"), full.names = TRUE, pattern = "large_cell_impacts_marine_fish-dominant")

fd_impacts_df <- lapply(fd_impacts_list, qread) %>%
  bind_rows()

qsave(fd_impacts_df, here("analysis/int/large_cell_impacts_marine_fish-dominant_regular_economic.qs"))

pd_impacts_list <- list.files(here("analysis/int/chunked/"), full.names = TRUE, pattern = "large_cell_impacts_marine_plant-dominant")

pd_impacts_df <- lapply(pd_impacts_list, qread) %>%
  bind_rows()

qsave(pd_impacts_df, here("analysis/int/large_cell_impacts_marine_plant-dominant_regular_economic.qs")) 


```


What IUCN category are these heavily impacted species? 

```{r}
# take a look at plant-dominant big impacts
max_impact_df_terrestrial_pd <- qread( here(this_dir, "int/large_cell_impacts_terrestrial_plant-dominant_regular_economic.qs"))

iucn_cats <- read.csv(here("prep/03_prep_spp_habitats/int/spp_iucn_cats.csv"))


max_impact_cell_spp <- max_impact_df_terrestrial_pd %>% 
  filter(cell_id == 786772) %>%
  mutate(species_full = tolower(species_full)) %>% 
  left_join(iucn_cats, by = c("species_full" = "sciname")) ## ok they are all LC lol. Majority of impacts from wheat and rapeseed

length(unique(max_impact_cell_spp$species_full)) # 39

## is it possible to see the iucn categories for those species which are not impacted in that cell? 


```

Ok, let's pull the "hotspot" areas for some examples:

 - Chile (coastal)
 - Canada
 - Norway
 - Argentina
 - Australia


```{r}
max_impact_df_terrestrial_pd <- qread(here(this_dir, "int/large_cell_impacts_terrestrial_plant-dominant_regular_economic.qs"))

max_impact_df_marine_pd <- qread( here(this_dir, "int/large_cell_impacts_marine_plant-dominant_regular_economic.qs"))

land_eez_rgns <- read.csv(file.path(biodiv_dir, "spatial/rgns_xy_all.csv")) %>%
  dplyr::select(x, y, iso3c) %>% 
  left_join(moll_template_xy)


max_impact_iso3c_terr <- max_impact_df_terrestrial_pd %>%
  left_join(land_eez_rgns) %>%
  dplyr::select(species = species_full, cell_id, impact_km2, taxon = spp_type, raw_material, allocation, fcr, diet, x, y, iso3c) %>%
  mutate(species = tolower(species)) %>%
    mutate(raw_material = case_when(raw_material == "CropsNES" ~ "Cropsnes",
                                    TRUE ~ raw_material))

unique(max_impact_iso3c_terr$raw_material) # [1] "Cropsnes"  "Maize"     "Pulses"    "Rapeseed"  "Soybean"   "Sunflower" "Wheat"  

max_impact_iso3c_marine <- max_impact_df_marine_pd %>%
  left_join(land_eez_rgns) %>%
    dplyr::select(species, cell_id, impact_km2, taxon, raw_material = fish_type, allocation, fcr, diet, x, y, iso3c) %>%
  mutate(raw_material = case_when(raw_material == "forage fish" ~ "Forage fish",
                                  raw_material == "trimmings fish" ~ "Trimmings fish",
       TRUE ~ raw_material))

unique(max_impact_iso3c_marine$raw_material) # [1] "Forage fish"    "Trimmings fish"

unique(max_impact_iso3c_terr$iso3c) # [1] "CAN" "NOR" "AUS" "ARG" "CHL" ok cool 
unique(max_impact_iso3c_marine$iso3c) # [1] "NOR" "AUS" "CHL" "ARG" "CAN" 

q95 <- 0.00006466858 # taken from the publication mapping script

```

**Figure 4. Impact hotspot map of central Chile, under the plant-dominant feed scenario.**
 > Figure 4. Impact hotspot map of central Chile, under the plant-dominant feed scenario. (A) Cells with average impacts greater than or equal to the 95th quantile. (B) The raw material which contributes most to the impact in each hotspot cell. (C) The proportion of impacted species in each cell which is at-risk of extinction. (D) Reference map, using Open Street Map (https://www.openstreetmap.org/). 

```{r}
# Now Chile

## note, sum of mean impacts in chile is 9.62481139705

max_impact <- max_impact_iso3c_terr %>%
  filter(iso3c == "CHL") %>% # lets check out Chile
  rbind(max_impact_iso3c_marine %>% filter(iso3c == "CHL")) %>%
  filter(y > -4250000) %>%
  filter(y < -3900000) %>%
  filter(x< -6000000)  # filter for a small part of the coast that has both marine and land large impacts. kind of near Valparaíso I think?

impact_rast <- max_impact %>%
  # filter(raw_material == "forage fish") %>%
  group_by(x, y) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  rast(., crs = moll_template) %>%
  project(., "EPSG:4326")

plot(impact_rast) # ok this is looking good now...


summary <- max_impact %>% 
  group_by(x, y, cell_id, species, raw_material) %>%
  summarise(impact_km2 = sum(impact_km2)) %>%
  ungroup() %>%
  group_by(x, y, cell_id, species) %>%
  mutate(total_spp_impact = sum(impact_km2, na.rm = TRUE)) %>%
  mutate(prop_total_spp_impact = total_spp_impact/100) %>%
  ungroup() %>%
  mutate(prop_raw_material_impact = impact_km2/total_spp_impact) %>%
  group_by(x, y, cell_id) %>%
  mutate(avg_impact = mean(prop_total_spp_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(iucn_cats, by = c("species" = "sciname"))

raw_mat_summary <- summary %>% 
  group_by(x, y, cell_id) %>%
  filter(prop_raw_material_impact == max(prop_raw_material_impact)) %>%
  dplyr::select(x, y, cell_id, dom_raw_material = raw_material) %>%
  ungroup() %>%
  distinct(x, y, cell_id, dom_raw_material)


summary_fin <- max_impact %>%
  group_by(x, y, cell_id, species) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impact = impact_km2/100) %>%
  group_by(x, y, cell_id) %>%
  summarise(prop_impact_mean = mean(prop_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(raw_mat_summary) %>%
    mutate(imp_category = case_when(
   prop_impact_mean < 0.0005 ~ "95-99",
    prop_impact_mean >= 0.0005 ~ "≥99"
  ))

# %>%
  #filter(prop_impact_mean >= q95) 


summary_iucn <- summary %>% 
  filter(cell_id %in% unique(summary_fin$cell_id))
  
iucn <- summary %>% 
  filter(cell_id %in% unique(summary_fin$cell_id)) %>%
  filter(!(category %in% c("NE", "DD", "LC")))

perc_threatened_impacts <- round((sum(iucn$impact_km2)/sum(summary_iucn$impact_km2))*100,1) # 6.6% of impacts shown are on threatened species? 


limits <- summary_fin %>%
  dplyr::select(x, y, prop_impact_mean) %>%
  rast(., crs = moll_template)

## now lets try plotting this:
country_sf <- countries_shp %>%
                          filter(sov_a3 == "CHL") %>%
  #   st_transform(crs(impact_rast)) %>%
  # st_crop(., st_bbox(impact_rast))
   st_crop(., st_bbox(limits))

summary_fin <- summary_fin %>%
  mutate(dom_raw_material = ifelse(dom_raw_material == "Cropsnes", "Other crops", dom_raw_material))

summary_fin$dom_raw_material <- factor(summary_fin$dom_raw_material, levels = c("Rapeseed", "Forage fish", "Wheat", "Pulses", "Other crops"))

ggplot() +
  geom_tile(data = summary_fin, aes(x = x, y = y, fill = dom_raw_material)) + 
  #   geom_sf(data = summary_fin_sf, aes(fill = dom_raw_material)) + 
       geom_sf(data = country_sf, fill = NA, colour = "black", lwd = 0.3)  +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
              #  legend.text = element_text(angle = 90, hjust = 1),  # Tilt legend labels
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = pal_mats, labels = c("Rapeseed", "Forage\nfish", "Wheat", "Pulses", "Other\ncrops")) + 
    labs(fill = "Highest impact raw material", 
         x = "", y = "",
                  tag = bquote(bold("B"))) +
    guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  )) 


ggsave(here("analysis/MS_plots/fig4_raw_material.png"), dpi = 300, bg = "white",  height = 16, width = 16, units = "cm")


summary_fin$imp_category <- factor(summary_fin$imp_category, levels = c("0-24", "25-49", "50-74", "75-94", "95-99", "≥99"))

ggplot() +
  geom_tile(data = summary_fin, aes(x = x, y = y, fill = imp_category)) + 
       geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3)  +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = quantile_pal) + 
    labs(fill = "Proportion of habitat impacted \nQuantiles",
         x = "", y = "",
                  tag = bquote(bold("A"))) + 
      guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  ))

ggsave(here("analysis/MS_plots/fig4_hotspot_quantile.png"), dpi = 300, bg = "white", height = 16, width = 16, units = "cm")



## it would be nice to see the proportion of species impacted in the worst category out of the total species in that cell: so number of impacted species in the worst category in the cell / total number of species assessed - so in this case we don't care about the category of the non-impacted species, right? Lets try that out..

iucn_plot_df <- summary_iucn %>%
  mutate(iucn_rank = case_when(category %in% c("NE") ~ 0,
                               category == "DD" ~ 1, 
                               category == "LC" ~ 2,
                               category == "NT" ~ 3,
                               category == "VU" ~ 4,
                               category == 'EN' ~ 5,
                               category == "CR" ~ 6)) %>%
  mutate(category_large = case_when(category %in% c("NE", "DD", "LC") ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk"),
         new_rank = case_when(category %in% c("NE", "DD", "LC") ~ 1,
                                    TRUE ~ 2)) %>%
  left_join(all_richness, by = "cell_id") %>%
  rename(sum_spp_assessed = sum_spp) %>%
  group_by(x, y, cell_id, category_large) %>%
  mutate(nspp_cats = n_distinct(species)) %>%
  ungroup() %>%
  group_by(x, y, cell_id) %>%
  mutate(total_spp_impacted = n_distinct(species),
         max_iucn_rank = max(new_rank, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_spp_worst_iucn = nspp_cats/total_spp_impacted,
             prop_spp_worst_iucn_assessed = nspp_cats/sum_spp_assessed) %>%
  group_by(x, y, cell_id) %>%
  filter(new_rank == max_iucn_rank) %>%
  summarise(max_iucn_rank = max(new_rank, na.rm = TRUE), 
            prop_at_risk_impacted = mean(prop_spp_worst_iucn, na.rm = TRUE),
            prop_at_risk_assessed = mean(prop_spp_worst_iucn_assessed, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(max_iucn_cat = case_when(max_iucn_rank == 1 ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk")) %>%
  mutate(prop_at_risk_impacted = ifelse(max_iucn_rank == 1, 0, prop_at_risk_impacted))


## this shows the worst IUCN rank of the IMPACTED species, with the shading showing the proportion of species which are impacted in that worst category out of the total species impacted

## so for example, one of the cells is CR with a value of 0.14. This means that 14% of the species in the cell are impacted and CR, while the remainder are not impacted

ggplot() +
     #annotation_map_tile(type = "cartolight", zoom = 8) +
  geom_tile(data = iucn_plot_df, aes(x = x, y = y, fill = prop_at_risk_impacted)) + 
  geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3)  +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
     scale_fill_gradientn(
    colors = c("#46A440", "#F8D422", "#781211"),
    values = c(0, 0.04, 1),  # Adjusting the values to handle the 0 case
    na.value = "#46A440"  # In case of any NAs, but your data summary doesn't show NAs
  ) +
    labs(fill = "Proportion of impacted species at-risk", 
         x = "", y = "", 
         tag = bquote(bold("C"))) +
  guides(fill = guide_colorbar(
    title.position = "top", 
    title.hjust = 0.5,
    barwidth = unit(6.244166667, "cm"),
    barheight = 1
  ))

ggsave(here("analysis/MS_plots/fig4_iucn_new.png"), dpi = 300, bg = "white",  height = 16, width = 16, units = "cm")


## Now save reference plot: 

map_plot <- ggplot() +
      geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
   annotation_map_tile(type = "cartolight", zoom = 8) +
        geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
   theme_minimal() +
   theme(
      legend.position = "left",
      legend.title = element_blank(),
      axis.title.x = element_blank(),
      axis.title.y = element_blank(),
      legend.key = element_rect(fill = NA),
      axis.text.x = element_text(angle = 60, hjust = 1),
      text = element_text(family = "Arial"),
      panel.border = element_rect(colour = "grey69", fill = NA, size = 3, linetype = "longdash")
   ) +
   labs(
      x = "", y = "",
      tag = bquote(bold("D"))
   )      +
     theme(
      plot.margin = margin(t = 0, r = 0, b = 70, l = 0)  # Adjust margins to align with the map
   )


ggsave(here("analysis/MS_plots/fig4_base_map.png"), plot = map_plot, dpi = 300, bg = "white", height = 16, width = 16, units = "cm")

```


```{r}
#### Calculate some stats about this area

test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 1)
nrow(test)/(nrow(iucn_plot_df)) # 0.5578035 of cells are not threatened

test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 2, prop_at_risk_impacted == 1)
nrow(test)/nrow(iucn_plot_df)

summary_distinct <- distinct(summary, cell_id, avg_impact)

sum(summary_distinct$avg_impact) # 1.720966

chl_marine <- max_impact_iso3c_marine %>%
  filter(iso3c == "CHL", diet == "plant-dominant")
sum(chl_marine$impact_km2) # 548.1677

chl_terr <- max_impact_iso3c_terr %>%
  filter(iso3c == "CHL", diet == "plant-dominant")
sum(chl_terr$impact_km2) # 5225.716

max_impact <- max_impact_iso3c_terr %>%
  filter(iso3c == "CHL") %>% # lets check out Chile
  rbind(max_impact_iso3c_marine %>% filter(iso3c == "CHL")) %>%
  filter(y > -4250000) %>%
  filter(y < -3900000) %>%
  filter(x< -6000000) 

sum(max_impact$impact_km2)/(sum(chl_terr$impact_km2) + sum(chl_marine$impact_km2)) # 0.4014775 - this hotspot represents 40% of all aquafeed impacts in Chile's land area and EEZ. 

346*100 # 34600 km2; Chile land and eez area is 2743128 km2 (wikipedia)
# 34600/2743128


raw_mat_totals <- raw_mat_summary %>%
  group_by(dom_raw_material) %>% 
  summarise(n_cells = n()) %>%
  ungroup()

145/346 # 0.4473008 of the cells are dominated by rapeseed impacts

test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 2) %>%
  left_join(raw_mat_summary) %>%
  filter(dom_raw_material == "Forage fish")

mean(test$prop_at_risk_impacted, na.rm = TRUE) # 0.06974141 forage fish - of the forage fish cells, only ~6.9% of species in the cells are at-risk

test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 2) 


test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 2) %>%
  left_join(raw_mat_summary) %>%
  group_by(dom_raw_material) %>%
  summarise(n()) %>%
  mutate(prop = `n()`/153)

```


Now look at a hotspot in Canada

```{r}
# Now Canada

max_impact <- max_impact_iso3c_terr %>%
  filter(iso3c %in% c("CAN")) %>% # lets check out CAN
  rbind(max_impact_iso3c_marine %>% filter(iso3c %in% c("CAN"))) %>%
  filter(x < -6000000) %>%
  filter(x > -7000000,
         y < 5500000)
## this region is called the Ontario Peninsula or Golden Horseshoe. It is one of the countries' leading soybean areas. Ontario accoutns for ~50-60% of soybean production in Canada and msot of it is from this area of Ontario.

impact_rast <- max_impact %>%
  group_by(x, y) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  rast(., crs = moll_template)

plot(impact_rast) # ok this is looking good now...


summary <- max_impact %>% 
  group_by(x, y, cell_id, species, raw_material) %>%
  summarise(impact_km2 = sum(impact_km2)) %>%
  ungroup() %>%
  group_by(x, y, cell_id, species) %>%
  mutate(total_spp_impact = sum(impact_km2, na.rm = TRUE)) %>%
  mutate(prop_total_spp_impact = total_spp_impact/100) %>%
  ungroup() %>%
  mutate(prop_raw_material_impact = impact_km2/total_spp_impact) %>%
  group_by(x, y, cell_id) %>%
  mutate(avg_impact = mean(prop_total_spp_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(iucn_cats, by = c("species" = "sciname"))

raw_mat_summary <- summary %>% 
  group_by(x, y, cell_id) %>%
  filter(prop_raw_material_impact == max(prop_raw_material_impact)) %>%
  dplyr::select(x, y, cell_id, dom_raw_material = raw_material) %>%
  ungroup() %>%
  distinct(x, y, cell_id, dom_raw_material)


summary_fin <- max_impact %>%
  group_by(x, y, cell_id, species) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impact = impact_km2/100) %>%
  group_by(x, y, cell_id) %>%
  summarise(prop_impact_mean = mean(prop_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(raw_mat_summary) %>%
    mutate(imp_category = case_when(
   prop_impact_mean < 0.0006 ~ "95-99",
    prop_impact_mean >= 0.0006 ~ "≥99"
  ))

summary_iucn <- summary %>% 
  filter(cell_id %in% unique(summary_fin$cell_id))
  
iucn <- summary %>% 
  filter(cell_id %in% unique(summary_fin$cell_id)) %>%
  filter(!(category %in% c("NE", "DD", "LC")))

perc_threatened_impacts <- round((sum(iucn$impact_km2)/sum(summary_iucn$impact_km2))*100,1) # 9% of impacts shown are on threatened species? 


limits <- summary_fin %>%
  dplyr::select(x, y, prop_impact_mean) %>%
  rast(., crs = moll_template)

## now lets try plotting this:
country_sf <- countries_shp %>%
                          filter(sov_a3 == "CAN") %>%
   st_crop(., st_bbox(limits))

# Filter USA and Canada geometries
can_usa <- countries_shp %>%
  filter(sov_a3 %in% c("CAN", "US1"))

# Find the shared border (intersection of boundaries)
border_us_can <- st_intersection(
  st_boundary(can_usa[can_usa$sov_a3 == "CAN", ]),
  st_boundary(can_usa[can_usa$sov_a3 == "US1", ])
) %>%
  st_crop(., st_bbox(limits))

summary_fin <- summary_fin %>%
  mutate(dom_raw_material = ifelse(dom_raw_material == "Cropsnes", "Other crops", dom_raw_material))

# summary_fin$dom_raw_material <- factor(summary_fin$dom_raw_material, levels = c("Rapeseed", "Forage fish", "Wheat", "Pulses"))
ggplot(border_us_can) +
  geom_sf(color = "black")

ggplot() +
  geom_tile(data = summary_fin, aes(x = x, y = y, fill = dom_raw_material)) + 
  geom_sf(data = border_us_can, color = "black", lwd = 0.3) +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
  panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = pal_mats) +
    labs(fill = "Highest impact raw material", 
         x = "", y = "",
                  tag = bquote(bold("B"))) +
    guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  )) 

ggsave(here("analysis/SI_plots/SI_hotspot_CAN_raw_material.png"), dpi = 300, bg = "white",  height = 16, width = 16, units = "cm")


summary_fin$imp_category <- factor(summary_fin$imp_category, levels = c("0-24", "25-49", "50-74", "75-94", "95-99", "≥99"))

ggplot() +
  geom_tile(data = summary_fin, aes(x = x, y = y, fill = imp_category)) + 
       geom_sf(data = border_us_can, fill = NA,  colour = "black", lwd = 0.3)  +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = quantile_pal) + 
    labs(fill = "Proportion of habitat impacted \nQuantiles",
         x = "", y = "",
                  tag = bquote(bold("A"))) + 
      guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  ))

ggsave(here("analysis/SI_plots/SI_hotspot_CAN_quantile.png"), dpi = 300, bg = "white", height = 16, width = 16, units = "cm")



## it would be nice to see the proportion of species impacted in the worst category out of the total species in that cell: so number of impacted species in the worst category in the cell / total number of species assessed - so in this case we don't care about the category of the non-impacted species, right? Lets try that out..

iucn_plot_df <- summary_iucn %>%
  mutate(iucn_rank = case_when(category %in% c("NE") ~ 0,
                               category == "DD" ~ 1, 
                               category == "LC" ~ 2,
                               category == "NT" ~ 3,
                               category == "VU" ~ 4,
                               category == 'EN' ~ 5,
                               category == "CR" ~ 6)) %>%
  mutate(category_large = case_when(category %in% c("NE", "DD", "LC") ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk"),
         new_rank = case_when(category %in% c("NE", "DD", "LC") ~ 1,
                                    TRUE ~ 2)) %>%
  left_join(all_richness, by = "cell_id") %>%
  rename(sum_spp_assessed = sum_spp) %>%
  group_by(x, y, cell_id, category_large) %>%
  mutate(nspp_cats = n_distinct(species)) %>%
  ungroup() %>%
  group_by(x, y, cell_id) %>%
  mutate(total_spp_impacted = n_distinct(species),
         max_iucn_rank = max(new_rank, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_spp_worst_iucn = nspp_cats/total_spp_impacted,
             prop_spp_worst_iucn_assessed = nspp_cats/sum_spp_assessed) %>%
  group_by(x, y, cell_id) %>%
  filter(new_rank == max_iucn_rank) %>%
  summarise(max_iucn_rank = max(new_rank, na.rm = TRUE), 
            prop_at_risk_impacted = mean(prop_spp_worst_iucn, na.rm = TRUE),
            prop_at_risk_assessed = mean(prop_spp_worst_iucn_assessed, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(max_iucn_cat = case_when(max_iucn_rank == 1 ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk")) %>%
  mutate(prop_at_risk_impacted = ifelse(max_iucn_rank == 1, 0, prop_at_risk_impacted))


## this shows the worst IUCN rank of the IMPACTED species, with the shading showing the proportion of species which are impacted in that worst category out of the total species impacted

## so for example, one of the cells is CR with a value of 0.14. This means that 14% of the species in the cell are impacted and CR, while the remainder are not impacted

ggplot() +
  geom_tile(data = iucn_plot_df, aes(x = x, y = y, fill = prop_at_risk_impacted)) + 
  geom_sf(data = border_us_can, fill = NA,  colour = "black", lwd = 0.3)  +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
     scale_fill_gradientn(
    colors = c("#46A440", "#F8D422", "#781211"),
    values = c(0, 0.04, 1),  # Adjusting the values to handle the 0 case
    na.value = "#46A440"  # In case of any NAs, but your data summary doesn't show NAs
  ) +
    labs(fill = "Proportion of impacted species at-risk", 
         x = "", y = "", 
         tag = bquote(bold("C"))) +
  guides(fill = guide_colorbar(
    title.position = "top", 
    title.hjust = 0.5,
    barwidth = unit(6.244166667, "cm"),
    barheight = 1
  ))

ggsave(here("analysis/SI_plots/SI_hotspot_CAN_iucn_new.png"), dpi = 300, bg = "white",  height = 16, width = 16, units = "cm")


## Now save reference plot: 

map_plot <- ggplot() +
      geom_sf(data = border_us_can, fill = NA,  colour = "black", lwd = 0.3) +
   annotation_map_tile(type = "cartolight", zoom = 8) +
        geom_sf(data = border_us_can, fill = NA,  colour = "black", lwd = 0.3) +
   theme_minimal() +
   theme(
      legend.position = "left",
      legend.title = element_blank(),
      axis.title.x = element_blank(),
      axis.title.y = element_blank(),
      legend.key = element_rect(fill = NA),
      axis.text.x = element_text(angle = 60, hjust = 1),
      text = element_text(family = "Arial"),
      panel.border = element_rect(colour = "grey69", fill = NA, size = 3, linetype = "longdash")
   ) +
   labs(
      x = "", y = "",
      tag = bquote(bold("D"))
   )      +
     theme(
      plot.margin = margin(t = 0, r = 0, b = 70, l = 0)  # Adjust margins to align with the map
   )


ggsave(here("analysis/SI_plots/SI_CAN_base_map.png"), plot = map_plot, dpi = 300, bg = "white", height = 16, width = 16, units = "cm")

```


```{r}
#### Calculate some stats about this area

test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 1)
7/975 # 0.7% of cells are not threatened
test <- iucn_plot_df %>% 
  filter(max_iucn_rank == 2, prop_at_risk_impacted == 1)
344/975 # 35% of cells are at risk and all species in those cells are at risk 

summary_distinct <- distinct(summary, cell_id, avg_impact)

sum(summary_distinct$avg_impact) # 2.905138

chl_marine <- max_impact_iso3c_marine %>%
  filter(iso3c == "CAN", diet == "plant-dominant")
sum(chl_marine$impact_km2) # 117.7947

chl_terr <- max_impact_iso3c_terr %>%
  filter(iso3c == "CAN", diet == "plant-dominant")
sum(chl_terr$impact_km2) # 46936.8

max_impact <- max_impact_iso3c_terr %>%
  filter(iso3c == "CAN") %>% # lets check out Chile
  rbind(max_impact_iso3c_marine %>% filter(iso3c == "CAN")) %>%
  filter(x < -6000000) %>%
  filter(x > -7000000,
         y < 5500000)

sum(max_impact$impact_km2)/(sum(chl_terr$impact_km2) + sum(chl_marine$impact_km2)) # 16% - this hotspot represents 16% of all aquafeed impacts in Canada's land area and EEZ. 

975*100 # 97500 km2; Canada land and eez area is 9900000 + 5599077 km2 (wikipedia)
# 97500/15499077 # 0.6%


raw_mat_totals <- raw_mat_summary %>%
  group_by(dom_raw_material) %>% 
  summarise(n_cells = n()) %>%
  ungroup()

# 836/975
# [1] 0.8574359 # soy bean

```

Do the same thing, but for Norway, and save it as a supplementary plot

```{r}
# filter for NOR

max_impact_nor <- max_impact_iso3c_terr %>%
  filter(iso3c == "NOR") %>% # lets check out norway.. cool only 100k cells
  rbind(max_impact_iso3c_marine %>% filter(iso3c == "NOR")) # wow ok the land impacts are way more concentrate in norway than ocean impacts

nor_rast <- max_impact_nor %>%
  group_by(x, y) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  rast(., crs = moll_template)


plot(nor_rast) # ok this is looking good now...


nor_summary <- max_impact_nor %>% 
  group_by(x, y, cell_id, species, raw_material) %>%
  summarise(impact_km2 = sum(impact_km2)) %>%
  ungroup() %>%
  group_by(x, y,cell_id, species) %>%
  mutate(total_spp_impact = sum(impact_km2, na.rm = TRUE)) %>%
  mutate(prop_total_spp_impact = total_spp_impact/100) %>%
  ungroup() %>%
  mutate(prop_raw_material_impact = impact_km2/total_spp_impact) %>%
  group_by(x, y, cell_id) %>%
  mutate(avg_impact = mean(prop_total_spp_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(iucn_cats, by = c("species" = "sciname")) %>%
  mutate(raw_material = ifelse(raw_material == "Cropsnes", "Other crops", raw_material))

nor_raw_mat_summary <- nor_summary %>% 
  group_by(x, y, cell_id) %>%
  filter(prop_raw_material_impact == max(prop_raw_material_impact)) %>%
  dplyr::select(x, y, cell_id, dom_raw_material = raw_material) %>%
  ungroup() %>%
  distinct(x, y, cell_id, dom_raw_material)

nor_summary_fin <- max_impact_nor %>%
  group_by(x, y, cell_id, species) %>%
  summarise(impact_km2 = sum(impact_km2, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_impact = impact_km2/100) %>%
  group_by(x, y, cell_id) %>%
  summarise(prop_impact_mean = mean(prop_impact, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(nor_raw_mat_summary) %>%
  filter(prop_impact_mean >= q95) %>% 
  filter(y <= 7295048) %>%
  mutate(imp_category = case_when(
   prop_impact_mean < 0.0005 ~ "95-99",
    prop_impact_mean >= 0.0005 ~ "≥99"
  ))


nor_limits <- nor_summary_fin %>%
  dplyr::select(x, y, prop_impact_mean) %>%
  rast(., crs = moll_template)

nor_rast <- nor_rast %>%
  crop(., nor_limits)


## now lets try plotting this:
country_sf <- countries_shp %>%
                          filter(sov_a3 == "NOR") %>%
  st_crop(., st_bbox(nor_rast))

ggplot() +
   geom_tile(data = nor_summary_fin, aes(x = x, y = y, fill = dom_raw_material)) + 
      geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
              #  legend.text = element_text(angle = 90, hjust = 1),  # Tilt legend labels
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = pal_mats, labels = c("Forage\nfish", "Other\ncrops", "Pulses", "Wheat")) + 
    labs(fill = "Highest impact raw material", 
         x = "", y = "",
                  tag = bquote(bold("B"))) +
    guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  ))

ggsave(here("analysis/SI_plots/extra_plots/CS_NOR_raw_material.png"), height = 8, width = 8, units = "in", bg = "white")

nor_summary_fin$imp_category <- factor(nor_summary_fin$imp_category, levels = c("0-24", "25-49", "50-74", "75-94", "95-99", "≥99"))

ggplot() +
  geom_tile(data = nor_summary_fin, aes(x = x, y = y, fill = imp_category)) + 
      geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
  scale_fill_manual(values = quantile_pal) + 
    labs(fill = "Proportion of habitat impacted \nQuantiles",
         x = "", y = "",
                  tag = bquote(bold("A"))) + 
      guides(fill = guide_legend(
    title.position = "top", 
    title.hjust = 0.5, 
    barheight = 1,
    label.position = "bottom",
    label.hjust = 0.5,
    label.vjust = 1,
    keywidth = unit(1.5, "cm")
  ))

ggsave(here("analysis/SI_plots/extra_plots/CS_NOR_hotspot_quantile.png"), height = 8, width = 8, units = "in", bg = "white")



nor_summary_iucn <- nor_summary %>% 
  filter(cell_id %in% unique(nor_summary_fin$cell_id))
  
nor_iucn <- nor_summary %>% 
  filter(cell_id %in% unique(nor_summary_fin$cell_id)) %>%
  filter(!(category %in% c("NE", "DD", "LC")))

nor_summary_iucn <- nor_summary %>% 
  filter(cell_id %in% unique(nor_summary_fin$cell_id))

iucn_plot_df <- nor_summary_iucn %>%
  mutate(iucn_rank = case_when(category %in% c("NE") ~ 0,
                               category == "DD" ~ 1, 
                               category == "LC" ~ 2,
                               category == "NT" ~ 3,
                               category == "VU" ~ 4,
                               category == 'EN' ~ 5,
                               category == "CR" ~ 6)) %>%
  mutate(category_large = case_when(category %in% c("NE", "DD", "LC") ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk"),
         new_rank = case_when(category %in% c("NE", "DD", "LC") ~ 1,
                                    TRUE ~ 2)) %>%
  left_join(all_richness, by = "cell_id") %>%
  rename(sum_spp_assessed = sum_spp) %>%
  group_by(x, y, cell_id, category_large) %>%
  mutate(nspp_cats = n_distinct(species)) %>%
  ungroup() %>%
  group_by(x, y, cell_id) %>%
  mutate(total_spp_impacted = n_distinct(species),
         max_iucn_rank = max(new_rank, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_spp_worst_iucn = nspp_cats/total_spp_impacted,
             prop_spp_worst_iucn_assessed = nspp_cats/sum_spp_assessed) %>%
  group_by(x, y, cell_id) %>%
  filter(new_rank == max_iucn_rank) %>%
  summarise(max_iucn_rank = max(new_rank, na.rm = TRUE), 
            prop_at_risk_impacted = mean(prop_spp_worst_iucn, na.rm = TRUE),
            prop_at_risk_assessed = mean(prop_spp_worst_iucn_assessed, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(max_iucn_cat = case_when(max_iucn_rank == 1 ~ "Not threatened or data deficient",
                                    TRUE ~ "At-risk")) %>%
  mutate(prop_at_risk_impacted = ifelse(max_iucn_rank == 1, 0, prop_at_risk_impacted))


ggplot() +
  geom_tile(data = iucn_plot_df, aes(x = x, y = y, fill = prop_at_risk_impacted)) + 
      geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
  theme_minimal() +
      theme(legend.position = "bottom", ## put legend on bottom
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.key = element_rect(fill = NA),
        axis.text.x = element_text(angle = 60, hjust = 1),
        text=element_text(family="Arial"),
panel.border = element_rect(colour = "grey69", fill=NA, size=3, linetype = "longdash")) +
     scale_fill_gradientn(
    colors = c("#46A440", "#F8D422", "#781211"),
    values = c(0, 0.04, 1),  # Adjusting the values to handle the 0 case
    na.value = "#46A440"  # In case of any NAs, but your data summary doesn't show NAs
  ) +
    labs(fill = "Proportion of impacted species at-risk", 
         x = "", y = "", 
         tag = bquote(bold("C"))) +
  guides(fill = guide_colorbar(
    title.position = "top", 
    title.hjust = 0.5,
    barwidth = unit(6.244166667, "cm"),
    barheight = 1
  ))

ggsave(here("analysis/SI_plots/extra_plots/CS_NOR_iucn.png"), height = 8, width = 8, units = "in", bg = "white")



map_plot <- ggplot() +
      geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
    geom_tile(data = iucn_plot_df, aes(x = x, y = y), fill = NA) + # make sure extent matches
   annotation_map_tile(type = "cartolight", zoom = 8) +
        geom_sf(data = country_sf, fill = NA,  colour = "black", lwd = 0.3) +
   theme_minimal() +
   theme(
      legend.position = "left",
      legend.title = element_blank(),
      axis.title.x = element_blank(),
      axis.title.y = element_blank(),
      legend.key = element_rect(fill = NA),
      axis.text.x = element_text(angle = 60, hjust = 1),
      text = element_text(family = "Arial"),
      panel.border = element_rect(colour = "grey69", fill = NA, size = 3, linetype = "longdash")
   ) +
   scale_fill_manual(values = pal_mats) +
   labs(
      x = "", y = "",
      tag = bquote(bold("D"))
   )      +
     theme(
      plot.margin = margin(t = 0, r = 0, b = 70, l = 0)  # Adjust margins to align with the map
   )

ggsave(here("analysis/SI_plots/extra_plots/NOR_basemap.png"), plot = map_plot, dpi = 300, bg = "white", height = 16, width = 16, units = "cm")


```
